### 2022-08-22 #4697 @editorialbot
[PRE REVIEW]: ShakeNBreak: Navigating the defect configurational landscape
**Submitting author:** @ireaml (<a href="http://orcid.org/0000-0001-7651-0814">Irea Mosquera-Lois</a>)
**Repository:** https://github.com/SMTG-UCL/ShakeNBreak
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
