+++ paper.bib
@@ -1,11 +1,11 @@
 @article{Arrigoni:2021,
-  	doi = {10.1038/s41524-021-00537-1},
   	Author = {{Arrigoni}, M. and {Madsen}, G.~K.~H.},
   	Journal = {Npj Computational Materials},
   	Title = {{Evolutionary computing and machine learning for discovering of low-energy defect configurations}},
   	Year = 2021,
     number = {1},
     volume = {7},
+	doi = {10.1038/s41524-021-00537-1},
 }
 @article{Pickard:2011,
 	year = {2011},
@@ -72,8 +72,8 @@
   }
 @article{castep,
 	author = {Stewart J. Clark and Matthew D. Segall and Chris J. Pickard and Phil J. Hasnip and Matt I. J. Probert and Keith Refson and Mike C. Payne},
-	doi = {doi:10.1524/zkri.220.5.567.65075},
-	url = {https://doi.org/10.1524/zkri.220.5.567.65075},
+	doi = {10.1524/zkri.220.5.567.65075},
+	url = {10.1524/zkri.220.5.567.65075},
 	title = {{First principles methods using CASTEP}},
 	journal = {Zeitschrift für Kristallographie - Crystalline Materials},
 	number = {5-6},
@@ -89,7 +89,7 @@
 	pages = {2175-2196},
 	year = {2009},
 	issn = {0010-4655},
-	doi = {https://doi.org/10.1016/j.cpc.2009.06.022},
+	doi = {10.1016/j.cpc.2009.06.022},
 	url = {https://www.sciencedirect.com/science/article/pii/S0010465509002033},
 	author = {Volker Blum and Ralf Gehrke and Felix Hanke and Paula Havu and Ville Havu and Xinguo Ren and Karsten Reuter and Matthias Scheffler},
 }
@@ -126,20 +126,20 @@
 }
 @article{Kavanagh:2022,
 	author = {Kavanagh, SR and Scanlon, DO and Walsh, A and Freysoldt, C},
-	doi = {10.1039/d2fd00043a},
 	journal = {Faraday Discussions},
 	title = {Impact of metastable defect structures on carrier recombination in solar cells},
 	url = {http://dx.doi.org/10.1039/d2fd00043a},
-	year = {2022}
+	year = {2022},
+	doi = {10.1039/d2fd00043a},
 }
 @misc{Mosquera-Lois:2022,
-  doi = {10.48550/ARXIV.2207.09862},
-  url = {https://arxiv.org/abs/2207.09862},
   author = {Mosquera-Lois, Irea and Kavanagh, Seán R. and Walsh, Aron and Scanlon, David O.},
   keywords = {Materials Science (cond-mat.mtrl-sci), FOS: Physical sciences, FOS: Physical sciences},
   title = {Identifying the ground state structures of point defects in solids},
   publisher = {arXiv},
   year = {2022},
+  doi = {10.48550/ARXIV.2207.09862},
+  url = {https://arxiv.org/abs/2207.09862},
 }
 @article{pycdt,
 	title = {{PyCDT: A Python toolkit for modeling point defects in semiconductors and insulators}},
@@ -148,7 +148,7 @@
 	pages = {165-179},
 	year = {2018},
 	issn = {0010-4655},
-	doi = {https://doi.org/10.1016/j.cpc.2018.01.004},
+	doi = {10.1016/j.cpc.2018.01.004},
 	url = {https://www.sciencedirect.com/science/article/pii/S0010465518300079},
[…]
