### 2022-08-23 3c90d6a
add description of `config.yaml` keywords in CLI generation tutorial
Former-commit-id: 348443000840d3c36df63d1b3228f8050ec36e37
