+++ docs/Generation.rst
@@ -7,7 +7,7 @@ For a single defect
 -------------------
 To generate the distorted structures for a specific defect in a range of charge states, we need to specify
 the structure for the bulk material (with ``--bulk`` flag), the defect structure (``--defect``) and the charge
-states (either with ``--min-charge`` and ``--max-charge`` to specify a range of charge states or with ``-charge`` 
+states (either with ``--min-charge`` and ``--max-charge`` to specify a range of charge states or with ``-charge``
 for a single charge state):
 
 .. code:: bash
@@ -21,8 +21,8 @@ we'll get a warning and we'll need to specify the defect site with the ``--defec
 
     $ snb-generate --bulk bulk_structure.cif --defect vac_1_Cd_POSCAR --min-charge -2 --max-charge 0 --defect-coords 0 0 0 --code VASP
 
-To specify additional distortion parameters, we can use a 
-`config.yaml <https://github.com/SMTG-UCL/ShakeNBreak/blob/main/input_files/example_generate_config.yaml>`_ 
+To specify additional distortion parameters, we can use a
+`config.yaml <https://github.com/SMTG-UCL/ShakeNBreak/blob/main/input_files/example_generate_config.yaml>`_
 file like the one below and use the ``--config`` flag to specify its path (i.e. ``snb-generate --config ./my_config.yaml``).
 A detailed description of all the parameters is available in the Python API section (:ref:`shakenbreak.input subsection <api_input>`).
 
@@ -31,21 +31,21 @@ A detailed description of all the parameters is available in the Python API sect
     # Example config.yaml file for snb-generate
 
     # General/Distortion section
-    oxidation_states: # If not specified, the code will determine them
+    oxidation_states:  # If not specified, the code will determine them
        Cd: 2
        Te: -2
-    distortion_increment: 0.1 # Bond distortion increment
+    distortion_increment: 0.1  # Bond distortion increment
     distorted_elements:  # (Default = None, distorts nearest neighbours)
         vac_Cd_1:
-            Cd # Distort Cd atoms near the Cd interstitial
+            Cd  # Distort Cd atoms near the Cd interstitial
 
     # Rattle section
-    stdev: 0.25 # Rattle standard deviation
+    stdev: 0.25  # Rattle standard deviation
     d_min: 2.25  # Displacements that place atoms closer than d_min are penalised. (Default = 80% of auto-determined bulk bond length)
     active_atoms: None  # Atoms to apply rattle displacement to. (Default = all atoms)
-    max_attempts: 5000
-    max_disp: 2.0
-    local_rattle: True # If True, rattle displacements will tail-off as we more away from the defect site
+    max_attempts: 5000  # Limit for how many attempted rattle moves are allowed a single atom; if this limit is reached an `Exception` is raised
+    max_disp: 2.0  # Rattle moves that yields a displacement larger than max_disp will always be rejected. Rarely occurs, mostly used as a safety net
+    local_rattle: True  # If True, rattle displacements will tail-off as we more away from the defect site
 
 To display additional information about the generated distortions we can set the ``--verbose`` flag.
 
@@ -102,7 +102,7 @@ Python API section (:ref:`shakenbreak.input subsection <api_input>`).
         defect_coords: [0.0, 0.0, 0.0]  # Fractional coords for vacancies!
     Int_Cd_2:
         charges: [0, +1, +2]
-        defect_index: -1
+        defect_index: -1  # Lattice site of the interstitial
 
     # Distortion section
     distortion_increment: 0.1 # Increment for distortion range
@@ -111,12 +111,12 @@ Python API section (:ref:`shakenbreak.input subsection <api_input>`).
             Cd # Distort Cd atoms near the Cd interstitial
 
     # Rattle section
-    stdev: 0.25
[…]
