### 2022-08-23 615f974
mention rattle kwargs in Distortion class docstring
