### 2022-08-23 1e22d96
mention rattle kwargs in Distortion class docstring
Former-commit-id: 615f9741efc9108fb337b26ebcff3675f4af0754
