+++ shakenbreak/vasp.py
@@ -86,6 +86,7 @@ def _import_psp():
 
 class _PotcarSingleMod(PotcarSingle):
     """Modified PotcarSingle class."""
+
     def __init__(self, *args, **kwargs):
         super(self.__class__, self).__init__(*args, **kwargs)
 
