+++ docs/Generation.rst
@@ -24,7 +24,8 @@ we'll get a warning and we'll need to specify the defect site with the ``--defec
 To specify additional distortion parameters, we can use a
 `config.yaml <https://github.com/SMTG-UCL/ShakeNBreak/blob/main/input_files/example_generate_config.yaml>`_
 file like the one below and use the ``--config`` flag to specify its path (i.e. ``snb-generate --config ./my_config.yaml``).
-A detailed description of all the parameters is available in the Python API section (:ref:`shakenbreak.input subsection <api_input>`).
+A detailed description of all the parameters is available in the Python API section
+(:ref:`shakenbreak.input.Distortions class <api_input>`).
 
 .. code:: yaml
 
@@ -89,7 +90,7 @@ the following directory structures will be parsed correctly:
 To specify the charge state range for each defect, as well as other optional arguments, we can use a
 `config.yaml <https://github.com/SMTG-UCL/ShakeNBreak/blob/main/input_files/example_generate_all_config.yaml>`_ file
 like the one below. A detailed description of all the parameters is available in the
-Python API section (:ref:`shakenbreak.input subsection <api_input>`).
+Python API section (:ref:`shakenbreak.input.Distortions class <api_input>`).
 
 .. code:: yaml
 
