+++ setup.py
@@ -44,7 +44,7 @@ setup(
     install_requires=[
         "numpy",
         "pymatgen",
-        "pymatgen-analysis-defects"
+        "pymatgen-analysis-defects",
         "matplotlib",
         "ase",
         "pandas",
