### 2022-08-23 9a74541
Add JOSS badge to docs
