### 2022-08-23 48c4147
Update pymatgen version in main (changes for compatibility with new pymatgen release and pymatgen-analysis-defects will be performed in develop)
Former-commit-id: 745a02c4c7c63fa30709b9ff34c1b0db244a6f51
