### 2022-08-23 149e94d
Remove duplicate args description in Distortions() docstring (convention to do this way I believe and means we don't get duplicate docstrings in tutorial notebook)
Former-commit-id: cac55f807bc21882b0a2a2be6e2625714ae7de48
