### 2022-08-23 2d60911
Add tutorial notebook to docs (lot of effort to make this look ok, I think... 😅)
