### 2022-08-23 ed84e73
Add tutorial notebook to docs (lot of effort to make this look ok, I think... 😅)
Former-commit-id: 2d60911c1eda531d8520011a757e7f5a846cc782
