### 2022-08-23 a00132f
Update module requirements for readthedocs
