+++ README.md
@@ -53,8 +53,7 @@ For development work, ShakeNBreak can also be installed from a copy of the sourc
 ## Usage
 
 ### Python API
-`ShakeNBreak` can be used through a Python API, as exemplified in the [Jupyter notebook](https://github.com/SMTG-UCL/ShakeNBreak/blob/main/tutorials/ShakeNBreak_Example_Workflow.ipynb), which can be run interactively by clicking here: [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/SMTG-UCL/ShakeNBreak/HEAD?filepath=tutorials)
-More info is given on the [docs](https://readthedocs.org/projects/shakenbreak) site.
+`ShakeNBreak` can be used through a Python API, as exemplified in the [SnB Python API tutorial](https://shakenbreak.readthedocs.io/en/latest/ShakeNBreak_Example_Workflow.html), with more info available on the [docs](https://readthedocs.org/projects/shakenbreak).
 
 ### Command line interface
 Alternatively, the code can be used via the command line. The functions provided include:
@@ -69,8 +68,7 @@ Alternatively, the code can be used via the command line. The functions provided
 
 More information about each function and its inputs/outputs are available from the [CLI section of the docs](https://shakenbreak.readthedocs.io/en/latest/shakenbreak.cli.html#commands) or using `-h` help option (e.g. `snb -h`).
 
-We recommend at least looking through the [tutorials](https://shakenbreak.readthedocs.io/en/latest/Tutorials.html) or the example Jupyter notebook ([![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/SMTG-UCL/ShakeNBreak/HEAD?filepath=tutorials)) when first starting to use `ShakeNBreak`, to familiarise yourself with the full functionality and workflow.
-
+We recommend at least looking through the [tutorials](https://shakenbreak.readthedocs.io/en/latest/Tutorials.html) when first starting to use `ShakeNBreak`, to familiarise yourself with the full functionality and workflow.
 ## Contributing
 
 ### Bugs reports, feature requests and questions
