+++ docs/requirements.txt
@@ -13,3 +13,4 @@ monty
 click
 shakenbreak
 doped
+importlib
