### 2022-08-24 e2d713e
Switch to importlib-metadata to avoid Module Error (see https://stackoverflow.com/questions/59216175/importerror-cannot-import-name-metadata-from-importlib and https://blog.finxter.com/fixed-modulenotfounderror-no-module-named-importlib-metadata/#:~:text=Quick%20Fix%3A%20Python%20raises%20the,with%20pip%20install%20importlib%2Dmetadata%20.)
