+++ docs/requirements.txt
@@ -13,4 +13,4 @@ monty
 click
 shakenbreak
 doped
-importlib
+importlib-metadata
