+++ docs/Tutorials.rst
@@ -5,7 +5,7 @@ Tutorials
 
 Python API
 ----------
-Usage of ShakeNBreak from a Python API is exemplified in the :ref:`SnB Python API tutorial <ShakeNBreak_Example_Workflow>`.
+Usage of ShakeNBreak from a Python API is exemplified in the `SnB Python API Tutorial <https://shakenbreak.readthedocs.io/en/latest/ShakeNBreak_Example_Workflow.html>`_.
 
 
 Command line interface (CLI)
