### 2022-08-24 e08a750
Update `setup.py` just so next release shows long PyPI description (currently empty) and links to docs
