### 2022-08-24 3666d20
Add note about complex defects to docs
Former-commit-id: f3a2b9bba6fe498aaa586aa21defcca0324504f5
