### 2022-08-24 c0d6827
Add note about metastable defects to docs
