+++ docs/Tips.rst
@@ -70,4 +70,12 @@ we can share some guidance for this (until improved ``pymatgen``-based functiona
 
 .. |chemsci| replace:: *Chem Sci* paper
 
+
+Metastable Defects
+----------
+While the ``ShakeNBreak`` workflow is primarily geared toward ground-state structure identification, it can also be
+applicable to finding metastable states, as described in the `method paper <https://arxiv.org/abs/2207.09862>`_.
+For this, you can use the optional ``metastable`` argument for ``get_energy_lowering_distortions``;
+see `docs here <https://shakenbreak.readthedocs.io/en/latest/shakenbreak.energy_lowering_distortions.html#shakenbreak.energy_lowering_distortions.get_energy_lowering_distortions>`_.
+
 Have any tips for users from using `ShakeNBreak`? Please share it with the developers and we'll add them here!
