### 2022-08-24 99667ad
Add catch with warning if all rattled distortions are >0.1 eV higher energy than unperturbed -> direct to docs site and suggest adjusting rattle std_dev
Former-commit-id: 2711e8a96cd07ce93446f34640ac33b10a9ab3bb
