+++ tests/test_energy_lowering_distortions.py
@@ -54,21 +54,15 @@ class EnergyLoweringDistortionsTestCase(unittest.TestCase):
             if not os.path.exists(f"{self.VASP_CDTE_DATA_DIR}/{defect_dir}"):
                 os.mkdir(self.VASP_CDTE_DATA_DIR + f"/{defect_dir}")
         # Int_Cd_2_1 without data, to test warnings
-        V_Cd_1_dict = {
-            "distortions": {-0.075: -205.700},
-            "Unperturbed": -205.800
-        }
+        V_Cd_1_dict = {"distortions": {-0.075: -205.740}, "Unperturbed": -205.800}
         dumpfn(
             V_Cd_1_dict,
-            os.path.join(self.VASP_CDTE_DATA_DIR, "vac_1_Cd_-1/vac_1_Cd_-1.yaml")
+            os.path.join(self.VASP_CDTE_DATA_DIR, "vac_1_Cd_-1/vac_1_Cd_-1.yaml"),
         )
-        V_Cd_2_dict = {
-            "distortions": {-0.35: -206.000},
-            "Unperturbed": -205.800
-        }
+        V_Cd_2_dict = {"distortions": {-0.35: -206.000}, "Unperturbed": -205.800}
         dumpfn(
             V_Cd_2_dict,
-            os.path.join(self.VASP_CDTE_DATA_DIR, "vac_1_Cd_-2/vac_1_Cd_-2.yaml")
+            os.path.join(self.VASP_CDTE_DATA_DIR, "vac_1_Cd_-2/vac_1_Cd_-2.yaml"),
         )
 
         self.defect_charges_dict = {
@@ -95,8 +89,7 @@ class EnergyLoweringDistortionsTestCase(unittest.TestCase):
             self.FHI_AIMS_DATA_DIR,
             self.ESPRESSO_DATA_DIR,
         ]:
-            for defect_dir in self.defect_folders_list + ["vac_1_Cd_-1",
-                                                          "vac_1_Cd_-2"]:
+            for defect_dir in self.defect_folders_list + ["vac_1_Cd_-1", "vac_1_Cd_-2"]:
                 if_present_rm(os.path.join(data_dir, defect_dir))
 
     def test_read_defects_directories(self):
@@ -198,7 +191,7 @@ class EnergyLoweringDistortionsTestCase(unittest.TestCase):
                 self.assertEqual(warning.category, UserWarning)
             self.assertIn(
                 f"Path {self.VASP_CDTE_DATA_DIR}/Int_Cd_2_1/Int_Cd_2_1.yaml does not exist",
-                str(w[0].message)
+                str(w[0].message),
             )
 
             self.assertEqual(len(low_energy_defects_dict), 1)
@@ -363,11 +356,11 @@ class EnergyLoweringDistortionsTestCase(unittest.TestCase):
         # was found for two different charge states
         V_Cd_1_dict_w_distortion = {
             "distortions": {-0.075: -206.700},
-            "Unperturbed": -205.800
+            "Unperturbed": -205.800,
         }
         dumpfn(
             V_Cd_1_dict_w_distortion,
-            os.path.join(self.VASP_CDTE_DATA_DIR, "vac_1_Cd_-1/vac_1_Cd_-1.yaml")
+            os.path.join(self.VASP_CDTE_DATA_DIR, "vac_1_Cd_-1/vac_1_Cd_-1.yaml"),
         )
         with patch("builtins.print") as mock_print:
             low_energy_defects_dict = (
@@ -494,6 +487,86 @@ class EnergyLoweringDistortionsTestCase(unittest.TestCase):
             )
         )
 
+    def test_get_energy_lowering_distortions_rattle_too_large(self):
[…]
