### 2022-08-24 f89c99d
Update `tearDown` for `energy_lowering_distortions` (other code tests were generating many files locally)
