+++ tests/test_energy_lowering_distortions.py
@@ -80,6 +80,19 @@ class EnergyLoweringDistortionsTestCase(unittest.TestCase):
             "sub_1_In_on_Cd_1",
         ]
 
+        self.orig_castep_0pt3_files = os.listdir(
+            self.CASTEP_DATA_DIR + "/vac_1_Cd_0/Bond_Distortion_30.0%"
+        )
+        self.orig_cp2k_0pt3_files = os.listdir(
+            self.CP2K_DATA_DIR + "/vac_1_Cd_0/Bond_Distortion_30.0%"
+        )
+        self.orig_fhi_aims_0pt3_files = os.listdir(
+            self.FHI_AIMS_DATA_DIR + "/vac_1_Cd_0/Bond_Distortion_30.0%"
+        )
+        self.orig_espresso_0pt3_files = os.listdir(
+            self.ESPRESSO_DATA_DIR + "/vac_1_Cd_0/Bond_Distortion_30.0%"
+        )
+
     def tearDown(self):
         # removed generated folders
         for data_dir in [
@@ -91,6 +104,31 @@ class EnergyLoweringDistortionsTestCase(unittest.TestCase):
         ]:
             for defect_dir in self.defect_folders_list + ["vac_1_Cd_-1", "vac_1_Cd_-2"]:
                 if_present_rm(os.path.join(data_dir, defect_dir))
+
+        # remove generated files
+        for data_dir, orig_files in [
+            (self.CASTEP_DATA_DIR, self.orig_castep_0pt3_files),
+            (self.CP2K_DATA_DIR, self.orig_cp2k_0pt3_files),
+            (self.FHI_AIMS_DATA_DIR, self.orig_fhi_aims_0pt3_files),
+            (self.ESPRESSO_DATA_DIR, self.orig_espresso_0pt3_files),
+        ]:  # everything but VASP data dir
+            if_present_rm(os.path.join(data_dir, "vac_1_Cd_0", "fake_vac_1_Cd_0.yaml"))
+            if_present_rm(os.path.join(data_dir, "vac_1_Cd_0", "vac_1_Cd_0.yaml"))
+            for defect_dir in os.listdir(os.path.join(data_dir, "vac_1_Cd_0")):
+                if "Bond_Distortion_30.0%" not in defect_dir and os.path.isdir(
+                    os.path.join(data_dir, "vac_1_Cd_0", defect_dir)
+                ):
+                    if_present_rm(os.path.join(data_dir, "vac_1_Cd_0", defect_dir))
+
+            for file in os.listdir(
+                os.path.join(data_dir, "vac_1_Cd_0", "Bond_Distortion_30.0%")
+            ):
+                if file not in orig_files:
+                    if_present_rm(
+                        os.path.join(
+                            data_dir, "vac_1_Cd_0", "Bond_Distortion_30.0%", file
+                        )
+                    )
 
     def test_read_defects_directories(self):
         """Test reading defect directories and parsing to dictionaries"""
@@ -551,7 +589,9 @@ class EnergyLoweringDistortionsTestCase(unittest.TestCase):
                 "not be included in low_energy_defects (check relaxation folders with CONTCARs "
                 "are present)."  # check this is skipped if no data
             )
-            self.assertEqual(len(w), 2)  # No Int_Cd_2_1 data and too large rattle warnings
+            self.assertEqual(
+                len(w), 2
+            )  # No Int_Cd_2_1 data and too large rattle warnings
             for warning in w:
[…]
