### 2022-08-25 93838e8
update docs conf file to show __init__ docstring with class docstring (otherwise args described in Distortion __init__ method not shown in docs)
