+++ docs/conf.py
@@ -110,6 +110,10 @@ intersphinx_mapping = {
     "matplotlib": ("http://matplotlib.org", None),
 }
 
+# -- Options for autodoc -----------------------------------------------------
+autoclass_content="both"
+
+# -- Options for nb extension -----------------------------------------------
 nb_execution_mode = "off"
 nb_render_image_options = {"height": "300",}  # Reduce plots size
 #myst_render_markdown_format = "gfm"
