### 2022-08-25 5e59758
update version of new release (to be compatible with pymatgen version v2022.7.25, while refactoring for compatibility with v2022.8.23 takes place)
Former-commit-id: 0182105795e9b5c19106792617687ca31c175a17
