### 2022-08-25 0182105
update version of new release (to be compatible with pymatgen version v2022.7.25, while refactoring for compatibility with v2022.8.23 takes place)
