### 2022-08-29 4e02fcd
Update plotting file naming as discussed (`vac_1_Cd_0` rather than `$V_{Cd}^{0}$` etc.)
