### 2022-08-29 e6e663f
Update plotting file naming as discussed (`vac_1_Cd_0` rather than `$V_{Cd}^{0}$` etc.)
Former-commit-id: 4e02fcda1b55674bb081ab500446409d337ed4d6
