+++ shakenbreak/plotting.py
@@ -1001,7 +1001,8 @@ def plot_defect(
             fig = plot_colorbar(
                 energies_dict=energies_dict,
                 disp_dict=disp_dict,
-                defect_name=defect_name,
+                defect_species=defect_species,
+                include_site_num_in_name=include_site_num_in_name,
                 title=defect_name if add_title else None,
                 num_nearest_neighbours=num_nearest_neighbours,
                 neighbour_atom=neighbour_atom,
@@ -1016,7 +1017,8 @@ def plot_defect(
         else:
             fig = plot_datasets(
                 datasets=[energies_dict],
-                defect_name=defect_name,
+                defect_species=defect_species,
+                include_site_num_in_name=include_site_num_in_name,
                 title=defect_name if add_title else None,
                 num_nearest_neighbours=num_nearest_neighbours,
                 neighbour_atom=neighbour_atom,
@@ -1032,7 +1034,8 @@ def plot_defect(
 def plot_colorbar(
     energies_dict: dict,
     disp_dict: dict,
-    defect_name: str,
+    defect_species: str,
+    include_site_num_in_name: Optional[bool] = False,
     num_nearest_neighbours: int = None,
     neighbour_atom: str = "NN",
     title: Optional[str] = None,
@@ -1056,9 +1059,14 @@ def plot_colorbar(
             Dictionary matching bond distortions to structure comparison metric
             (metric = 'disp' or 'max_dist'), as produced by
             `analysis.calculate_struct_comparison()`.
-        defect_name (:obj:`str`):
-            Specific defect name that will appear in plot labels and file names
-            (e.g '$V_{Cd}^0$')
+        defect_species (:obj:`str`):
+            Specific defect name that will appear in plot labels (in LaTeX form)
+             and file names (e.g 'vac_1_Cd_0')
+        include_site_num_in_name (:obj:`bool`):
+            Whether to include the site number (as generated by doped) in the
+            defect name. Useful for materials with many symmetry-inequivalent
+            sites.
+            (Default: False)
         num_nearest_neighbours (:obj:`int`):
             Number of distorted neighbour atoms (e.g. 2)
             (Default: None)
@@ -1107,7 +1115,8 @@ def plot_colorbar(
         ax = _format_axis(
             ax=ax,
             y_label=y_label,
-            defect_name=defect_name,
+            defect_name=_format_defect_name(defect_species,
+                                            include_site_num_in_name=include_site_num_in_name),
             num_nearest_neighbours=num_nearest_neighbours,
             neighbour_atom=neighbour_atom,
         )
@@ -1211,8 +1220,8 @@ def plot_colorbar(
         # Plot reference energy
         unperturbed_color = colormap(
             0
-        )  # get color of unperturbed structure (corresponding to 0 as disp is calculated with respect
-        # to this structure)
+        )  # get color of unperturbed structure (corresponding to 0 as disp is calculated with
+        # respect to this structure)
         ax.scatter(
             0,
             energies_dict["Unperturbed"],
@@ -1244,7 +1253,7 @@ def plot_colorbar(
     if save_plot:
         _save_plot(
             fig=fig,
-            defect_name=defect_name,
+            defect_name=defect_species,
             save_format=save_format,
         )
     return fig
@@ -1253,7 +1262,8 @@ def plot_colorbar(
 def plot_datasets(
     datasets: list,
     dataset_labels: list,
-    defect_name: str,
+    defect_species: str,
+    include_site_num_in_name: Optional[bool] = False,
     title: Optional[str] = None,
     neighbour_atom: Optional[str] = None,
     num_nearest_neighbours: Optional[int] = None,
@@ -1277,9 +1287,14 @@ def plot_datasets(
[…]
