### 2022-08-29 7844200
Update CLI run script to suppress cp warning message when alternative cp command succeeds
Former-commit-id: b8716d2aac41a7f03bb38f808d83a4c7aa60b6c2
