+++ shakenbreak/bash_scripts/SnB_run.sh
@@ -45,7 +45,7 @@ SnB_run_loop () {
         then
         builtin cd "$i" || return
         if [ ! -f "${job_filename}" ] && [ ! "$job_in_cwd" = false ]
-          then "cp" ../../"${job_filename}" . || "cp" ../"${job_filename}" . || return
+          then "cp" ../../"${job_filename}" . 2>/dev/null || "cp" ../"${job_filename}" . || return
         fi
         if [ -f OUTCAR ]  # if OUTCAR exists so rerunning rather than 1st run
           then echo "${i%?} not (fully) relaxed, saving files and rerunning"
