### 2022-08-29 fad63fa
Update `setup.py` to include `input_files` as package data, so `pip install`ed package includes these data files
