### 2022-08-29 c634844
Refactor to `SnB_input_files` to avoid confusion or potential overwriting in users' `site-packages` folder
Former-commit-id: 35df5436d0579db0ec9857b899cd700534f4181a
