### 2022-08-29 35df543
Refactor to `SnB_input_files` to avoid confusion or potential overwriting in users' `site-packages` folder
