### 2022-08-29 522a8f2
Create pip_install_test.yml
