+++ .github/workflows/pip_install_test.yml
@@ -1,0 +1,41 @@
+name: pip install ShakeNBreak
+
+on:
+  pull_request:
+    branches:
+       - main  # only test on main as develop updates not on PyPI
+  push:
+    branches:
+      - main
+  workflow_dispatch:
+
+jobs:
+  build-linux:
+    runs-on: ubuntu-latest
+    strategy:
+      max-parallel: 5
+
+      matrix:
+        python-version: ['3.8', '3.9']  # Excluding 3.10 for the moment, as hiphive requires Numpy <1.22 which is incompatible with Python 3.10, should be fixed soon...
+
+    name: Python ${{ matrix.python-version }} pip install
+
+    steps:
+      - uses: actions/checkout@v2
+
+      - uses: actions/setup-python@v2
+        with:
+          python-version: ${{ matrix.python-version }}
+
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install shakenbreak  # install only from PyPI
+
+      - name: Test
+        run:  |
+          pytest --ignore=tests/test_local.py  # Ignore local tests file, which tests INCAR and POTCAR file writing but not possible on GitHub Actions
+          pytest --mpl tests/test_shakenbreak.py  # test output plots
+          pytest --mpl tests/test_plotting.py  # test output plots
+          # pytest --mpl-generate-path=tests/remote_baseline tests/test_plotting.py  # generate output plots
+          # pytest --mpl-generate-path=tests/remote_baseline tests/test_shakenbreak.py  # generate output plots
