### 2022-08-29 3908641
Add [tests] to pip install
