+++ .github/workflows/pip_install_test.yml
@@ -31,6 +31,7 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install shakenbreak  # install only from PyPI
+          pip install shakenbreak[tests]
 
       - name: Test
         run:  |
