### 2022-08-29 5cb6a77
Update `snb-parse` help message to inform of default naming (current directory name) when not specified
