+++ shakenbreak/cli.py
@@ -974,7 +974,8 @@ def run(submit_command, job_script, job_name_option, all, verbose):
 @click.option(
     "--defect",
     "-d",
-    help="Name of defect (including charge state) to parse energies for (e.g. 'vac_1_Cd_0')",
+    help="Name of defect (including charge state) to parse energies for (e.g. 'vac_1_Cd_0'). "
+         "Default is current directory name.",
     type=str,
     default=None,
 )
