### 2022-08-30 71821b4
Update LCHARG and LWAVE in default INCAR as discussed
