### 2022-08-30 cceceea
Update EDIFFG in default INCAR as discussed
Former-commit-id: 20b077c6c8bb1b89f244c236187ccee6e92beac0
