### 2022-08-30 cd94f67
Remove DefectSet.yaml as discussed
Former-commit-id: 1e2b465cec2422f584e6ffaae5f9c21df66533da
