### 2022-08-30 743d1f5
Remove `prepare_vasp_defect_inputs()` because inherited from `doped` (I think) but not required and only used in one test? (Now refactored to no longer require this)
Former-commit-id: 9942fe6ba7ca71595d57636d221c4d6faf24d87e
