### 2022-08-30 9942fe6
Remove `prepare_vasp_defect_inputs()` because inherited from `doped` (I think) but not required and only used in one test? (Now refactored to no longer require this)
