### 2022-08-30 da37af3
Initial commits for adding compatibility to pymatgen>2022.8.23, after breaking changes to defects module
Former-commit-id: 7600db0dfe091900655c881803c55a00101d5d8d
