### 2022-08-31 652d005
update style sheet and test for local plots
