### 2022-08-31 46f9350
update setup to install shakenbreak custom font
Former-commit-id: 0054b678f9dd8899419feffea950ce8371200a8b
