### 2022-08-31 fa57e19
Remove `LVHAR` from default `INCAR` as corrections not required or typically done for distortion tests -> save file space
