### 2022-08-31 bc2c8c2
Set LREAL = Auto in `incar.yaml` (think this was a typo setting to False?)
Former-commit-id: cf5de34b118c99429ae9ebf7b5d436effd303580
