### 2022-08-31 145011b
Set mpl backend to cairo if save_format = pdf, and add pycairo as dependency
