+++ tests/test_local.py
@@ -208,7 +208,9 @@ class DistortionLocalTestCase(unittest.TestCase):
             k: v for k, v in vasp.default_incar_settings.items() if "#" not in k
         }  # pymatgen doesn't parsed commented lines
         self.parsed_incar_settings_wo_comments = {
-            k: v for k, v in self.parsed_default_incar_settings.items() if "#" not in str(v)
+            k: v
+            for k, v in self.parsed_default_incar_settings.items()
+            if "#" not in str(v)
         }  # pymatgen ignores comments after values
 
         with open(
@@ -235,7 +237,12 @@ class DistortionLocalTestCase(unittest.TestCase):
         for i in os.listdir("."):
             if "distortion_metadata" in i:
                 os.remove(i)
-            elif "Vac_Cd" in i or "Int_Cd" in i or "Wally_McDoodle" in i or "pesky_defects" in i:
+            elif (
+                "Vac_Cd" in i
+                or "Int_Cd" in i
+                or "Wally_McDoodle" in i
+                or "pesky_defects" in i
+            ):
                 shutil.rmtree(i)
 
     # test create_folder and create_vasp_input simultaneously:
@@ -400,7 +407,8 @@ class DistortionLocalTestCase(unittest.TestCase):
         parsed_settings = self.parsed_incar_settings_wo_comments.copy()
         parsed_settings.pop("ENCUT")
         self.assertTrue(
-            parsed_settings.items() <= V_Cd_INCAR.items()  # matches after
+            parsed_settings.items()
+            <= V_Cd_INCAR.items()  # matches after
             # removing kwarg settings
         )
         V_Cd_KPOINTS = Kpoints.from_file(V_Cd_minus50_folder + "/KPOINTS")
@@ -443,14 +451,15 @@ class DistortionLocalTestCase(unittest.TestCase):
         # Test the following options:
         # --defect, --path, --format,  --units, --colorbar, --metric, --title, --verbose
         defect = "vac_1_Ti_0"
-        wd = os.getcwd()  # plots saved to distortion_plots directory in current directory
+        wd = (
+            os.getcwd()
+        )  # plots saved to distortion_plots directory in current directory
         dumpfn(
             {
-                "distortions":
-                    {-0.4: -1176.28458753},
+                "distortions": {-0.4: -1176.28458753},
                 "Unperturbed": -1173.02056574,
             },
-            f"{self.EXAMPLE_RESULTS}/{defect}/{defect}.yaml"
+            f"{self.EXAMPLE_RESULTS}/{defect}/{defect}.yaml",
         )
         if os.path.exists(f"{self.EXAMPLE_RESULTS}/distortion_metadata.json"):
             os.remove(f"{self.EXAMPLE_RESULTS}/distortion_metadata.json")
@@ -471,7 +480,7 @@ class DistortionLocalTestCase(unittest.TestCase):
                     "--colorbar",
                     "--metric",
                     "disp",
-                    "-t", # No title
+                    "-t",  # No title
                     "-v",
                 ],
                 catch_exceptions=False,
@@ -479,11 +488,15 @@ class DistortionLocalTestCase(unittest.TestCase):
         self.assertTrue(os.path.exists(wd + "/distortion_plots/vac_1_Ti_0.png"))
         compare_images(
             wd + "/distortion_plots/vac_1_Ti_0.png",
-            f"{file_path}/local_baseline_plots/"+"vac_1_Ti_0_cli_colorbar_disp.png",
+            f"{file_path}/local_baseline_plots/" + "vac_1_Ti_0_cli_colorbar_disp.png",
             tol=2.0,
[…]
