### 2022-08-31 5f49dc2
refactor to install custom font when importing plotting module (if font hasn't been already installed)
Former-commit-id: 9e902e241a714339dea0f940d9cd2f1bee1eea9e
