+++ setup.py
@@ -2,37 +2,39 @@
 """
 This is a setup.py script to install ShakeNBreak
 """
+from distutils.command import install_headers
 import os
 from setuptools import setup, find_packages
 from setuptools.command.install import install
+from setuptools.command.develop import develop
+from setuptools.command.egg_info import egg_info
+from distutils.command.build_py import build_py as _build_py
+from distutils.cmd import Command
 import warnings
 
+path_to_file = os.path.dirname(os.path.abspath(__file__))
 
-# Set up the machinery to install custom fonts.  Subclass the setup tools install
-# class in order to run custom commands during installation.
+
 # See https://stackoverflow.com/questions/34193900/how-do-i-distribute-fonts-with-my-python-package
-class move_ttf(install):
-    def run(self):
-        """
-        Performs the usual install process and then copies the True Type fonts
-        that come with SnB into matplotlib's True Type font directory,
-        and deletes the matplotlib fontList.cache.
-        """
-        # Perform the usual install process
-        install.run(self)
-        print("Trying to install Whitney font...")
-        # Try to install custom font
+def _install_custom_font():
+    """Install ShakeNBreak custom font."""
+    print("Trying to install ShakeNBreak custom font...")
+    # Try to install custom font
+    try:
         try:
             import os, shutil
             import matplotlib as mpl
             import matplotlib.font_manager
+        except:
+            print("Cannot import matplotlib!")
 
-            # Find where matplotlib stores its True Type fonts
-            mpl_data_dir = os.path.dirname(mpl.matplotlib_fname())
-            mpl_fonts_dir = os.path.join(mpl_data_dir, 'fonts', 'ttf')
+        # Find where matplotlib stores its True Type fonts
+        mpl_data_dir = os.path.dirname(mpl.matplotlib_fname())
+        mpl_fonts_dir = os.path.join(mpl_data_dir, 'fonts', 'ttf')
 
-            # Copy the font file to matplotlib's True Type font directory
-            fonts_dir = "fonts/"
+        # Copy the font file to matplotlib's True Type font directory
+        fonts_dir = f"{path_to_file}/fonts/"
+        try:
             for file_name in os.listdir(fonts_dir):
                 if '.ttf' in file_name:  # must be in ttf format for matplotlib
                     old_path = os.path.join(fonts_dir, file_name)
@@ -41,29 +43,62 @@ class move_ttf(install):
                     print("Copying " + old_path + " -> " + new_path)
                 else:
                     print(f"No ttf fonts found in the {fonts_dir} directory.")
-
-            # Try to delete matplotlib's fontList cache
-            mpl_cache_dir = mpl.get_cachedir()
-            mpl_cache_dir_ls = os.listdir(mpl_cache_dir)
-            if 'fontList.cache' in mpl_cache_dir_ls:
-                fontList_path = os.path.join(mpl_cache_dir, 'fontList.cache')
-                if fontList_path:
-                    os.remove(fontList_path)
-                    print("Deleted the matplotlib fontList.cache.")
-            else:
-                print("Couldn't find matplotlib cache, so will continue.")
-
-            # Add fonts
-            for font in os.listdir(fonts_dir):
-                print(font)
-                matplotlib.font_manager._load_fontmanager(try_read_cache=False)
-                matplotlib.font_manager.fontManager.addfont(f"{fonts_dir}/{font}")
-                print("Adding Whitney font to matplotlib fonts.")
         except:
-            warnings.warn(
-                "WARNING: An issue occured while installing the custom font for ShakeNBreak. "
-                "The widely available Helvetica font will be used instead."
-            )
[…]
