### 2022-08-31 5c8a03e
Refactor `_setup_distorted_defect_dict()` to just take 1 rather than 3 arguments
