### 2022-08-31 a8aa76f
Change `pycairo` to optional extra, and add code + warning to determine backend for pdf figures
Former-commit-id: 0858f2d199ac376d4044abed318d969f2002a096
