### 2022-08-31 3a86a2c
update test plots with custom font
Former-commit-id: efabdd259457970a49f752a97090ebf2ca58f61b
