### 2022-08-31 efabdd2
update test plots with custom font
