+++ .github/workflows/build_and_test.yml
@@ -45,13 +45,13 @@ jobs:
           # pytest --mpl-generate-path=tests/remote_baseline tests/test_plotting.py  # generate output plots
           # pytest --mpl-generate-path=tests/remote_baseline tests/test_shakenbreak.py  # generate output plots
 
-      - name: Archive test plots
-        uses: actions/upload-artifact@v3
-        with:
-          name: output-plots
-          path: tests/remote_baseline
+      # - name: Archive test plots
+      #   uses: actions/upload-artifact@v3
+      #   with:
+      #     name: output-plots
+      #     path: tests/remote_baseline
 
-      - name: Download a single artifact
-        uses: actions/download-artifact@v3
-        with:
-          name: output-plots
+      # - name: Download a single artifact
+      #   uses: actions/download-artifact@v3
+      #   with:
+      #     name: output-plots
