+++ shakenbreak/plotting.py
@@ -2,7 +2,8 @@
 Module containing functions to plot distorted defect relaxation outputs and identify
 energy-lowering distortions.
 """
-import os, shutil
+import os
+import shutil
 import warnings
 from typing import Optional, Tuple
 import numpy as np
@@ -545,6 +546,7 @@ def _save_plot(
     if "pdf" in save_format:
         try:
             import cairo
+
             backend = "cairo"
         except ImportError:
             warnings.warn(
