+++ docs/conf.py
@@ -23,7 +23,7 @@ copyright = '2022, Irea Mosquera-Lois, Seán R. Kavanagh'
 author = 'Irea Mosquera-Lois, Seán R. Kavanagh'
 
 # The full version, including alpha/beta/rc tags
-release = '1.0.2'
+release = '22.9.1'
 
 
 # -- General configuration ---------------------------------------------------
