### 2022-09-01 cf5dffb
Refactor snb-analyse and snb-plot to also now be able to run from a single defect directory (if no arguments given – simplifies CLI workflow even more, nice for CLI video & docs etc), like snb-run and snb-parse. Also updating the warnings
Former-commit-id: 511c807dfe2e9f043908741f9f7efffdb77e8ff7
