+++ shakenbreak/cli.py
@@ -943,8 +943,9 @@ def generate_all(
 )
 def run(submit_command, job_script, job_name_option, all, verbose):
     """
-    Loop through distortion subfolders for a defect (or for all defect folders in the current
-    directory, if the --all (-a) flag is set) and submit calculations to the HPC scheduler.
+    Loop through distortion subfolders for a defect, when run within a defect folder, or for all
+    defect folders in the current (top-level) directory if the --all (-a) flag is set, and submit
+    jobs to the HPC scheduler.
     """
     optional_flags = "-"
     if all:
@@ -969,13 +970,14 @@ def run(submit_command, job_script, job_name_option, all, verbose):
 @snb.command(
     name="parse",
     context_settings=CONTEXT_SETTINGS,
-    no_args_is_help=False,  # to allow convenient parsing
+    no_args_is_help=False,  # can be run within defect directory with no options/arguments set
 )
 @click.option(
     "--defect",
     "-d",
-    help="Name of defect (including charge state) to parse energies for (e.g. 'vac_1_Cd_0'). "
-    "Default is current directory name.",
+    help="Name of defect species (folder) to parse (e.g. 'vac_1_Cd_0'), if run from "
+         "top-level directory or above. Default is current directory name (assumes running from "
+         "within defect folder).",
     type=str,
     default=None,
 )
@@ -1017,32 +1019,34 @@ def parse(defect, all, path, code):
     else:
         # assume current directory is the defect folder
         try:
-            if path == "." or path == "./":
-                path = os.getcwd()
-            if path.endswith("/"):
-                defect = path.split("/")[-2]
-                path = path.rsplit("/", 2)[0]
-            else:
-                defect = path.split("/")[-1]
-                path = path.rsplit("/", 1)[0]
+            if path != ".":
+                warnings.warn(
+                    "`--path` option ignored when running from within defect folder (i.e. "
+                    "when `--defect` is not specified.")
+            path = os.getcwd()
+            defect = path.split("/")[-1]
+            path = path.rsplit("/", 1)[0]
             io.parse_energies(defect, path, code)
-        except Exception:
-            warnings.warn(
-                "Could not parse energies! Please specify a defect to parse "
-                "(with option --defect) or use the --all flag to parse all defects"
-                " present in the specified/current directory."
+        except:
+            raise Exception(
+                f"Could not parse defect '{defect}' in directory '{path}'. Please either specify "
+                f"a defect to parse (with option --defect), run from within a single defect "
+                f"directory (without setting --defect) or use the --all flag to parse all "
+                f"defects in the specified/current directory."
             )
 
 
 @snb.command(
     name="analyse",
     context_settings=CONTEXT_SETTINGS,
-    no_args_is_help=True,
+    no_args_is_help=False,  # can be run within defect directory with no options/arguments set
 )
 @click.option(
     "--defect",
     "-d",
-    help="Name of defect (including charge state) to analyse (e.g. 'vac_1_Cd_0')",
+    help="Name of defect species (folder) to analyse and plot (e.g. 'vac_1_Cd_0'), if run from "
+         "top-level directory or above. Default is current directory name (assumes running from "
+         "within defect folder).",
     type=str,
     default=None,
 )
@@ -1118,38 +1122,50 @@ def analyse(defect, all, path, code, ref_struct, verbose):
         for defect in defect_dirs:
             print(f"\nAnalysing {defect}...")
             analyse_single_defect(defect, path, code, ref_struct, verbose)
-    elif defect:
-        # Check if defect present in path, in case path not specified and user is
-        # analysing defect from inside the defect directory
-        if path == ".":
+    elif defect is None:
+        # assume current directory is the defect folder
+            if path != ".":
[…]
