### 2022-09-01 4929833
Update tests in response to CLI functions update
Former-commit-id: b7d6672b6bb7ddca2ac3b96d4a27787eb6d514b3
