### 2022-09-01 a9f0b62
Add snb-analyse within defect folder (no args) test
Former-commit-id: c7abb65a09cd3fef83d030ddb7a349bd9f26ac1d
