### 2022-09-01 b8d0ad6
Add test for exception when running snb-analyse above defect folder but without args
Former-commit-id: a84faac1fe1141c3f351faaa20ded61388374305
