### 2022-09-01 9c87365
Add snb-parse tests for warning when run without setting defect, but setting path
