+++ tests/test_cli.py
@@ -1333,6 +1333,40 @@ nonsense_key: nonsense_value"""
             )
         )
         os.chdir(file_path)
+
+        # Test warning when setting path and parsing from inside the defect folder
+        defect_name = "vac_1_Ti_-1"
+        os.remove(
+            f"{self.EXAMPLE_RESULTS}/pesky_defects/{defect_name}/{defect_name}.yaml"
+        )
+        os.chdir(f"{self.EXAMPLE_RESULTS}/pesky_defects/{defect_name}")
+        with warnings.catch_warnings(record=True) as w:
+            result = runner.invoke(
+                snb,
+                [
+                    "parse",
+                    "-p",
+                    self.EXAMPLE_RESULTS,
+                ],
+                catch_exceptions=False,
+            )
+        self.assertTrue(any([warning.category == UserWarning for warning in w]))
+        self.assertTrue(
+            any(
+                [
+                    str(warning.message)
+                    == "`--path` option ignored when running from within defect folder (i.e. "
+                    "when `--defect` is not specified."
+                    for warning in w
+                ]
+            )
+        )
+        self.assertTrue(
+            os.path.exists(
+                f"{self.EXAMPLE_RESULTS}/pesky_defects/{defect_name}/{defect_name}.yaml"
+            )
+        )
+        os.chdir(file_path)
         shutil.rmtree(f"{self.EXAMPLE_RESULTS}/pesky_defects/")
 
     def test_parse_codes(self):
@@ -1536,7 +1570,7 @@ nonsense_key: nonsense_value"""
 
         # Test analysing from inside the defect folder
         defect_name = "vac_1_Ti_0"
-        os.chdir(f"{self.EXAMPLE_RESULTS}/{defect_name}")
+        os.chdir(self.VASP_TIO2_DATA_DIR)
         with warnings.catch_warnings(record=True) as w:
             result = runner.invoke(
                 snb,
@@ -1545,12 +1579,21 @@ nonsense_key: nonsense_value"""
                 ],
                 catch_exceptions=True,
             )
-        self.assertTrue([warning.category == UserWarning for warning in w])
-        self.assertTrue([str(warning.message) == "No output file in Bond_Distortion_10.0% "
-                                                 "directory" for warning in w])
+        self.assertTrue(any([warning.category == UserWarning for warning in w]))
+        self.assertTrue(
+            any(
+                [
+                    str(warning.message) == "No output file in Bond_Distortion_10.0% "
+                    "directory"
+                    for warning in w
+                ]
+            )
+        )
         self.assertIn("Comparing structures to Unperturbed...", result.output)
-        self.assertIn(f"Saved results to {os.path.join(os.getcwd(), defect_name)}.csv",
-                      result.output)
+        self.assertIn(
+            f"Saved results to {os.path.join(os.getcwd(), defect_name)}.csv",
+            result.output,
+        )
         self.assertTrue(os.path.exists(f"{defect_name}.csv"))
         self.assertTrue(os.path.exists(f"{defect_name}.yaml"))
         os.remove(f"{defect_name}.csv")
@@ -1568,8 +1611,13 @@ nonsense_key: nonsense_value"""
             str(result.exception),
         )
         self.assertNotIn(f"Saved results to", result.output)
-        self.assertFalse(any(os.path.exists(i) for i in os.listdir() if (i.endswith(".csv") or
-                                                                         i.endswith(".yaml"))))
+        self.assertFalse(
+            any(
+                os.path.exists(i)
+                for i in os.listdir()
+                if (i.endswith(".csv") or i.endswith(".yaml"))
+            )
+        )
         os.chdir(file_path)
 
     def test_plot(self):
