### 2022-09-01 d7ab864
Add warning to io.parse_energies when no results parsed, and add test for this in test_cli.py (snb-parse when run with no arguments in top-level folder)
