+++ shakenbreak/io.py
@@ -243,6 +243,11 @@ def parse_energies(
         if energies != {"distortions": {}}:
             energies = sort_energies(energies)
             save_file(energies, defect, path)
+        else:
+            warnings.warn(f"Energies could not be parsed for defect '{defect}' in {path}. "
+                          f"If these directories are correct, check calculations have converged, "
+                          f"and that distortion subfolders match ShakeNBreak naming (e.g. "
+                          f"Bond_Distortion_xxx, Rattled, Unperturbed)")
 
 
 # Parsing output structures of different codes
