### 2022-09-01 4ad37f7
Add warning to io.parse_energies when no results parsed, and add test for this in test_cli.py (snb-parse when run with no arguments in top-level folder)
Former-commit-id: d7ab86464c886531ef02df5320f15466ca8f3d8b
