### 2022-09-01 8010d8f
Add snb-analyse tests for warning when run without setting defect, but setting path
