+++ .github/workflows/build_and_test.yml
@@ -39,19 +39,19 @@ jobs:
 
       - name: Test
         run:  |
-          # pytest --ignore=tests/test_local.py  # Ignore local tests file, which tests INCAR and POTCAR file writing but not possible on GitHub Actions
-          # pytest --mpl tests/test_shakenbreak.py  # test output plots
-          # pytest --mpl tests/test_plotting.py  # test output plots
-          pytest --mpl-generate-path=tests/remote_baseline tests/test_plotting.py  # generate output plots
-          pytest --mpl-generate-path=tests/remote_baseline tests/test_shakenbreak.py  # generate output plots
+          pytest --ignore=tests/test_local.py  # Ignore local tests file, which tests INCAR and POTCAR file writing but not possible on GitHub Actions
+          pytest --mpl tests/test_shakenbreak.py  # test output plots
+          pytest --mpl tests/test_plotting.py  # test output plots
+          # pytest --mpl-generate-path=tests/remote_baseline tests/test_plotting.py  # generate output plots
+          # pytest --mpl-generate-path=tests/remote_baseline tests/test_shakenbreak.py  # generate output plots
 
-      - name: Archive test plots
-        uses: actions/upload-artifact@v3
-        with:
-          name: output-plots
-          path: tests/remote_baseline
+      # - name: Archive test plots
+      #   uses: actions/upload-artifact@v3
+      #   with:
+      #     name: output-plots
+      #     path: tests/remote_baseline
 
-      - name: Download a single artifact
-        uses: actions/download-artifact@v3
-        with:
-          name: output-plots
+      # - name: Download a single artifact
+      #   uses: actions/download-artifact@v3
+      #   with:
+      #     name: output-plots
