### 2022-09-01 30421d0
update test plots
Former-commit-id: 65fae1abdadcc04a600db5ec7b44c4f42f68b189
