### 2022-09-01 332cd89
Add snb-plot test for running in defect folder with no arguments
