### 2022-09-01 d042af4
Add snb-plot tests for warning when run without setting defect, but setting path for running in defect folder with no arguments, and test for exception when run in top-level directory with no arguments
Former-commit-id: 308b20a141bd6d7d92d169b865d94d9010526392
