### 2022-09-01 ec1f2c5
Update `write_groundstate_structure()` to also work for a single defect folder
