### 2022-09-01 e1d8b0a
Add direct structure comparison tests to `test_groundstate()`
