### 2022-09-01 d9079af
Update snb-groundstate to allow running with no args within a defect folder (automatically determines if in defect folder or top-level directory). Update docstrings, and add tests for this (plus warning when run while setting `--path` option)
