+++ shakenbreak/cli.py
@@ -12,7 +12,7 @@ import fnmatch
 from monty.serialization import loadfn
 from monty.json import MontyDecoder
 
-from pymatgen.core.structure import Structure
+from pymatgen.core.structure import Structure, Element
 from pymatgen.io.vasp.inputs import Incar
 from pymatgen.analysis.defects.core import Defect
 
@@ -980,8 +980,8 @@ def run(submit_command, job_script, job_name_option, all, verbose):
     "--defect",
     "-d",
     help="Name of defect species (folder) to parse (e.g. 'vac_1_Cd_0'), if run from "
-         "top-level directory or above. Default is current directory name (assumes running from "
-         "within defect folder).",
+    "top-level directory or above. Default is current directory name (assumes running from "
+    "within defect folder).",
     type=str,
     default=None,
 )
@@ -1026,10 +1026,11 @@ def parse(defect, all, path, code):
             if path != ".":
                 warnings.warn(
                     "`--path` option ignored when running from within defect folder (i.e. "
-                    "when `--defect` is not specified.")
-            path = os.getcwd()
-            defect = path.split("/")[-1]
-            path = path.rsplit("/", 1)[0]
+                    "when `--defect` is not specified."
+                )
+            cwd = os.getcwd()
+            defect = cwd.split("/")[-1]
+            path = cwd.rsplit("/", 1)[0]
             io.parse_energies(defect, path, code)
         except:
             raise Exception(
@@ -1049,8 +1050,8 @@ def parse(defect, all, path, code):
     "--defect",
     "-d",
     help="Name of defect species (folder) to analyse and plot (e.g. 'vac_1_Cd_0'), if run from "
-         "top-level directory or above. Default is current directory name (assumes running from "
-         "within defect folder).",
+    "top-level directory or above. Default is current directory name (assumes running from "
+    "within defect folder).",
     type=str,
     default=None,
 )
@@ -1128,12 +1129,14 @@ def analyse(defect, all, path, code, ref_struct, verbose):
             analyse_single_defect(defect, path, code, ref_struct, verbose)
     elif defect is None:
         # assume current directory is the defect folder
-            if path != ".":
-                warnings.warn("`--path` option ignored when running from within defect folder ("
-                              "i.e. when `--defect` is not specified.")
-            path = os.getcwd()
-            defect = path.split("/")[-1]
-            path = path.rsplit("/", 1)[0]
+        if path != ".":
+            warnings.warn(
+                "`--path` option ignored when running from within defect folder ("
+                "i.e. when `--defect` is not specified."
+            )
+        cwd = os.getcwd()
+        defect = cwd.split("/")[-1]
+        path = cwd.rsplit("/", 1)[0]
 
     defect = defect.strip("/")  # Remove trailing slash if present
     # Check if defect present in path:
@@ -1161,8 +1164,8 @@ def analyse(defect, all, path, code, ref_struct, verbose):
     "--defect",
     "-d",
     help="Name of defect species (folder) to analyse and plot (e.g. 'vac_1_Cd_0'), if run from "
-         "top-level directory or above. Default is current directory name (assumes running from "
-         "within defect folder).",
+    "top-level directory or above. Default is current directory name (assumes running from "
+    "within defect folder).",
     type=str,
     default=None,
 )
@@ -1286,12 +1289,14 @@ def plot(
         )
     elif defect is None:
         # assume current directory is the defect folder
-            if path != ".":
-                warnings.warn("`--path` option ignored when running from within defect folder ("
-                              "i.e. when `--defect` is not specified.")
-            path = os.getcwd()
-            defect = path.split("/")[-1]
-            path = path.rsplit("/", 1)[0]
+        if path != ".":
+            warnings.warn(
[…]
