### 2022-09-02 15299da
Remove print statement (for mpl_dir) from `_install_custom_font()` (for cleaner output etc.). Think this was for testing/development purposes
Former-commit-id: c2acbff596d05434609d24b034644b0facc4989a
