### 2022-09-02 be01fcc
Move `_install_custom_font()`to beginnging of `plot_colorbar` and `plot_datasets` to avoid redundant calls
Former-commit-id: f475101717286b4698ba53eb64fabe1f689bfac3
