### 2022-09-02 6b9e0ef
Fix colorbar typo
Former-commit-id: 4f9875fefd1c07f8c20b41b32c4fa85586e7677d
