### 2022-09-02 65aa36a
Explicitly set plot x-lims to avoid missing datapoints
Former-commit-id: 91b078250aeb84f21ee7b374f495b09ec037cf53
