### 2022-09-02 9c6c8e2
Update missing `distortion_metadata.json` warning so user know what difference this makes
Former-commit-id: 7dbcad23c2b09c98b41be4d4613df779974e0f6e
