### 2022-09-02 044a796
Refactor `--title` to `--no-title` as discussed
