### 2022-09-02 bf24891
Update CLI tests
