### 2022-09-02 7d599d5
Avoid io warning about no outputs parsed if results present in previously parsed yaml file (and use these results if so)
