### 2022-09-02 0f13641
Set `local_rattle = False` in `test_local.py` so tests pass
