### 2022-09-02 41dcb30
Update local example plots and docs plots to match new font
Former-commit-id: 4f6e00353b50a234e8e955e648cb98aaf9e7d6c4
