### 2022-09-02 8ff3205
Update CHANGELOG.rst before merging
Former-commit-id: 454a63b36855e68a065651e40c23dc036f8728f5
