### 2022-09-02 454a63b
Update CHANGELOG.rst before merging
