### 2022-09-02 e6de7f8
Update `package_data` in `setup.py` so fonts also installed from `pip install shakenbreak`
