### 2022-09-14 f71e284
Update tests and add test for `snb-run` with job script in different directory
Former-commit-id: 7230b38a29dc4aea479876352e240d637f69061c
