### 2022-09-14 1ca6da3
Move `charged_defect = {}` to within loop over defect charges, so that `Rattled` directory isn't also written for non-fully-ionised defects
