### 2022-09-14 67ee1d4
Add tests that Rattled not written for non-fully-ionised defects, and only Rattled and Unperturbed written for fully-ionised defects
