### 2022-09-15 9ab88dc
Add SnB summary GIF to readme and docs
Former-commit-id: 7bea89004c75c14e8325d4e4f6878e8ab959c902
