### 2022-09-15 4a00c7c
Require Click version 8.0 or higher, 7.x gives nargs error.
