### 2022-09-15 497f048
Update format defect name to handle all lower/uppercase defect naming (pymatgen default is upppercase "Vac" which `plotting` wasn't recognising)
