### 2022-09-15 3cda1e2
Update format defect name to handle all lower/uppercase defect naming (pymatgen default is upppercase "Vac" which `plotting` wasn't recognising)
Former-commit-id: 497f048008f8969596315944434a2a7b4e47276a
