### 2022-09-19 f679a96
Update oxidation state handling to (1) also handle cases where we have just extrinsic interstitials (and not substitutions) and (2) cases where explicit oxidation states are only provided for some of the species (in which case guess the rest and inform user)
