+++ shakenbreak/input.py
@@ -823,37 +823,62 @@ class Distortions:
         self.stdev = stdev
         self.local_rattle = local_rattle
 
-        if oxidation_states is None:
-            if "bulk" in self.defects_dict:
-                bulk_comp = self.defects_dict["bulk"]["supercell"][
-                    "structure"
-                ].composition
-                self.oxidation_states = bulk_comp.oxi_state_guesses()[0]
+        # check if all expected oxidation states are provided
+        if "bulk" in self.defects_dict:
+            bulk_comp = self.defects_dict["bulk"]["supercell"]["structure"].composition
+            guessed_oxidation_states = bulk_comp.oxi_state_guesses()[0]
 
-            else:  # determine bulk composition from first defect in dict
-                defect_subdict = list(self.defects_dict.values())[0][0]
-                bulk_comp = _get_bulk_comp(defect_subdict)
-                self.oxidation_states = bulk_comp.oxi_state_guesses()[0]
+        else:  # determine bulk composition from first defect in dict
+            defect_subdict = list(self.defects_dict.values())[0][0]
+            bulk_comp = _get_bulk_comp(defect_subdict)
+            guessed_oxidation_states = bulk_comp.oxi_state_guesses()[0]
 
-            if "substitutions" in self.defects_dict:
-                for substitution in self.defects_dict["substitutions"]:
-                    if (
-                        substitution["defect_type"] == "substitution"
-                        and substitution["bulk_supercell_site"].specie.symbol
-                        not in self.oxidation_states
-                    ):
-                        # substituting species not in bulk composition
-                        substitution_specie = substitution["substitution_specie"]
-                        likely_substitution_oxi = _most_common_oxi(substitution_specie)
-                        self.oxidation_states[
-                            substitution_specie
-                        ] = likely_substitution_oxi
+        if "substitutions" in self.defects_dict:
+            for substitution in self.defects_dict["substitutions"]:
+                if (
+                    substitution["defect_type"] == "substitution"
+                    and substitution["bulk_supercell_site"].specie.symbol
+                    not in guessed_oxidation_states
+                ):
+                    # extrinsic substituting species not in bulk composition
+                    extrinsic_specie = substitution["bulk_supercell_site"].specie
+                    likely_substitution_oxi = _most_common_oxi(extrinsic_specie)
+                    guessed_oxidation_states[
+                        extrinsic_specie
+                    ] = likely_substitution_oxi
 
-            print(
+        if "interstitials" in self.defects_dict:
+            for interstitial in self.defects_dict["interstitials"]:
+                if (
+                    interstitial["bulk_supercell_site"].specie.symbol
+                    not in guessed_oxidation_states
+                ):
+                    # extrinsic species not in bulk composition
+                    extrinsic_specie = interstitial["bulk_supercell_site"].specie
+                    likely_substitution_oxi = _most_common_oxi(extrinsic_specie)
+                    guessed_oxidation_states[
+                        extrinsic_specie
+                    ] = likely_substitution_oxi
+
+        if not guessed_oxidation_states.keys() <= self.oxidation_states.keys():
+            # some or all oxidation states are missing, so use guessed versions for these and
+            # inform user
+            if self.oxidation_states == None:
+                print(
                 f"Oxidation states were not explicitly set, thus have been "
                 f"guessed as {self.oxidation_states}. If this is unreasonable "
                 f"you should manually set oxidation_states"
             )
+                self.oxidation_states = guessed_oxidation_states
+
+            else:  # only some oxidation states are missing, so use guessed versions for these
+                # and inform user
+                missing_oxidation_states = guessed_oxidation_states.keys() - self.oxidation_states.keys()
[…]
