### 2022-09-19 309820d
Update tests to match this behaviour
Former-commit-id: 9964bcc7f4d3e0ae373d9011ca18a80de5bf4f8a
