### 2022-09-19 9964bcc
Update tests to match this behaviour
