### 2022-09-19 73b7844
Add tests for partial oxidation state specification, and that (weird) bulk/extrinsic oxidation states don't get overwritten when explicitly specified.
Former-commit-id: fd1b6b0fbc8f2dbf9a0bc19f00cb8bfbda674621
