### 2022-09-19 e33a107
Test print statements for oxidation states not set or partially set
