### 2022-09-19 6481fc3
Test extrinsic interstitial (only) defect and no print statement when all oxidation states set
