+++ shakenbreak/plotting.py
@@ -14,6 +14,8 @@ import matplotlib.pyplot as plt
 from matplotlib.figure import Figure
 import seaborn as sns
 
+from pymatgen.core.periodic_table import Element
+
 from shakenbreak import analysis
 
 MODULE_DIR = os.path.dirname(os.path.abspath(__file__))
@@ -160,6 +162,194 @@ def _format_defect_name(
     # Format defect name for title/axis labels
     if charge > 0:
         charge = "+" + str(charge)  # show positive charges with a + sign
+
+    defect_name = None
+    dummy_h = Element("H")
+    # Vacancy: v/vac/va/vacancy, underscore and element symbol
+    if any(
+            [
+                f"{substring}_" in defect_species.lower()  # defect shorthand plus underscore
+                and any((dummy_h.is_valid_symbol(subsubstring[:2]) or dummy_h.is_valid_symbol(
+                    subsubstring[-2:]) or dummy_h.is_valid_symbol(subsubstring[:1]) or
+                                           dummy_h.is_valid_symbol(
+                    subsubstring[-1:]) for subsubstring in defect_species.split("_") if
+                        subsubstring != substring))  #
+                # Element
+                # symbol right before or after an underscore
+                for substring in (
+                    "vac",
+                    "v",
+                    "va",
+                    "vacancy",
+            )  # TODO: Test each of these!
+            ]
+    ): # TODO: use this function whenever we test the defect name (search for 'subsstring'
+        # usages and "vac" usages)
+        # vacancy... defect_name
+
+    # else check if possibly different-format vacancy: (vCd/vaCd/vacCd/vacancyCd):
+    elif any( ## I feel like this could give a lot of false positives; Only test 2-char elements
+            # after "v"?
+            [
+                f"{substring}" in defect_species.lower()  # defect shorthand
+                and any((dummy_h.is_valid_symbol(subsubstring[:2]) or dummy_h.is_valid_symbol(
+                    subsubstring[:1]) for subsubstring in defect_species.split(substring) if
+                        subsubstring != substring))  # Element symbol right after v/va/vac/vacancy
+                for substring in (
+                    "vac",
+                    "v",
+                    "va",
+                    "vacancy")
+            ]
+    ):
+        # vacancy... poss_defect_name
+
+    # Interstitial: int/interstitial (case-insensitive) or i_ (case-sensitive), underscore and
+    # element for interstitial
+    if any(
+            [
+                (f"{substring}_" in defect_species.lower()  # defect shorthand plus underscore
+                # Element symbol right before or after an underscore
+                for substring in (
+                    "int",
+                    "interstitial",
+            ) or defect_species.startswith("i")) and any((dummy_h.is_valid_symbol(subsubstring[:2])
+                                                        or
+                                dummy_h.is_valid_symbol(
+                    subsubstring[-2:]) or dummy_h.is_valid_symbol(subsubstring[:1]) or
+                         dummy_h.is_valid_symbol(
+                             subsubstring[-1:]) for subsubstring in defect_species.split("_")))  # TODO: Test each of these!
+            ]
+    ):
+        # interstitial... defect_name
+        # if defect_name is not None: whoops
+
+    # or (new pymatgen naming) Element + _i (i.e. Cd_i)
+    if any(
+            [
+                f"_i" in defect_species  # defect shorthand plus underscore
+                and any(dummy_h.is_valid_symbol(defect_species.split("_i")[0][-2:]) or dummy_h.is_valid_symbol(
+                    defect_species.split("_i")[0][-1:]))  # Element symbol right before or after an underscor
+            ]
+    ):
+        # interstitial... defect_name
+        # if defect_name is not None: whoops
+
+    # else check if possibly different-format interstitial: (iCd/intCd/interstitialCd):
[…]
