### 2022-09-19 716f633
Set defaults: `EDIFFG = -0.01` and `local_rattle = False`
Former-commit-id: 3ca73b32586417aa331a3c57aece8e2274b39a2f
