+++ shakenbreak/input.py
@@ -742,7 +742,7 @@ class Distortions:
         dict_number_electrons_user: Optional[dict] = None,
         distortion_increment: float = 0.1,
         bond_distortions: Optional[list] = None,
-        local_rattle: bool = True,
+        local_rattle: bool = False,
         stdev: float = 0.25,
         distorted_elements: Optional[dict] = None,
         **kwargs,  # for mc rattle
@@ -776,10 +776,10 @@ class Distortions:
                 (Default: None)
             local_rattle (:obj:`bool`):
                 Whether to apply random displacements that tail off as we move
-                away from the defect site. Recommended as it is often faster than
-                the full rattle (requires less ionic relaxation steps). If False,
-                all supercell sites are rattled with the same amplitude (full ratlle).
-                (Default: True)
+                away from the defect site. Not recommended as typically worsens
+                performance. If False (default), all supercell sites are rattled
+                with the same amplitude (full rattle).
+                (Default: False)
             stdev (:obj:`float`):
                 Standard deviation (in Angstroms) of the Gaussian distribution
                 from which random atomic displacement distances are drawn during
