### 2022-09-19 af03350
Rollback unintentional _format_defect_name() update (still WIP)
