### 2022-09-19 e268060
Update tests to new SnB defaults
Former-commit-id: df2f676468c3c2d32fac24e8fda9d6e69d3bd9d0
