### 2022-09-19 df2f676
Update tests to new SnB defaults
