### 2022-09-19 6b56f69
Formatting for flake8, and ignore E203 (see https://black.readthedocs.io/en/stable/the_black_code_style/current_style.html#slices)
Former-commit-id: 5129656aa0ed4c19dbcf766fea57bf81f28aa705
