### 2022-09-19 43692ce
Make `path`/`defect` parsing more robust (i.e. don't fail if `defect` is present earlier in `path`, not in final bottom directory)
