### 2022-09-19 1c0b937
Add tests for when `defect` is present higher up in path, for `snb-plot` and `snb-analyse`
