### 2022-09-19 81f0e80
Move `if energies else warn` code outside `if defect_dir` condition, so if defect_dir isn't round the warning is printed (previously `snb-parse` would run without error/warning but nothing parse). Also handle if `defect` included in `path` as with `snb-plot` and `snb-analyse`
