### 2022-09-19 4f1a158
Add tests for io.parse_energies warning when nothing parsed because defect folder not recognised, and for handling when defect is present higher up in path
Former-commit-id: 7001402c34674d28361c8a47b4ce3662d329f2c2
