### 2022-09-19 7001402
Add tests for io.parse_energies warning when nothing parsed because defect folder not recognised, and for handling when defect is present higher up in path
