### 2022-09-20 4f2d484
Fix `snb-generate-all` test failure
Former-commit-id: 00b82925f50bbb78f6b10600fa7b15c276bf01b5
