### 2022-09-20 f3c8f6f
Update plotting behaviour and add tests
Former-commit-id: 068c12a459d5bd16c21218c100cc760dd9e7bca3
