### 2022-09-20 068c12a
Update plotting behaviour and add tests
