### 2022-09-20 9678c51
Add tests for _save_plot() expected behaviour (renaming, printing info etc)
