+++ tests/test_plotting.py
@@ -5,6 +5,7 @@ import shutil
 import warnings
 from copy import deepcopy
 import pytest
+import datetime
 
 import numpy as np
 from monty.serialization import loadfn, dumpfn
@@ -24,11 +25,12 @@ def if_present_rm(path):
             shutil.rmtree(path)
 
 
-file_path = os.path.dirname(__file__)
+_file_path = os.path.dirname(__file__)
 
 
 class PlottingDefectsTestCase(unittest.TestCase):
     def setUp(self):
+
         self.DATA_DIR = os.path.join(os.path.dirname(__file__), "data")
         self.VASP_CDTE_DATA_DIR = os.path.join(self.DATA_DIR, "vasp/CdTe")
         self.organized_V_Cd_distortion_data = loadfn(
@@ -479,7 +481,9 @@ class PlottingDefectsTestCase(unittest.TestCase):
         # Saving to defect_dir subfolder in output_path
         fig, ax = plt.subplots(1, 1)
         defect_name = "vac_1_Cd_0"
-        with patch("builtins.print") as mock_print, warnings.catch_warnings(record=True) as w:
+        with patch("builtins.print") as mock_print, warnings.catch_warnings(
+            record=True
+        ) as w:
             plotting._save_plot(
                 fig=fig,
                 defect_name=defect_name,
@@ -506,15 +510,79 @@ class PlottingDefectsTestCase(unittest.TestCase):
         mock_print.assert_called_once_with(f"Plot saved to ./vac_1_Cd_0.svg")
         if_present_rm(f"./{defect_name}.svg")
 
+        # Saving to defect_dir subfolder in output_path where output_path is cwd
+        os.chdir(self.VASP_CDTE_DATA_DIR)
+        with patch("builtins.print") as mock_print:
+            plotting._save_plot(
+                fig=fig, defect_name=defect_name, output_path=".", save_format="svg"
+            )
+        self.assertFalse(
+            os.path.exists(f"./{defect_name}.svg")
+        )  # not in cwd, in defect directory
+        self.assertTrue(os.path.exists(f"{defect_name}/{defect_name}.svg"))
+        mock_print.assert_called_once_with(
+            f"Plot saved to {defect_name}/{defect_name}.svg"
+        )
+        if_present_rm(f"{defect_name}/{defect_name}.svg")
+        os.chdir(_file_path)
 
-        # Saving to defect_dir subfolder in output_path where output_path is not cwd
-        # test old saved plot renaming and print statement
+        # test previously saved plot renaming and print statement
+        plotting._save_plot(
+            fig=fig, defect_name=defect_name, output_path=".", save_format="svg"
+        )
+        self.assertTrue(os.path.exists(f"./{defect_name}.svg"))
+        with patch("builtins.print") as mock_print:
+            plotting._save_plot(
+                fig=fig, defect_name=defect_name, output_path=".", save_format="svg"
+            )
+        current_datetime = datetime.datetime.now().strftime("%Y-%m-%d-%H-%M")
+        current_datetime_minus1min = (
+            datetime.datetime.now() - datetime.timedelta(minutes=1)
+        ).strftime(
+            "%Y-%m-%d-%H-%M"
+        )  # in case delay between writing and testing plot generation
+
+        self.assertTrue(os.path.exists(f"./{defect_name}.svg"))
+        mock_print.assert_any_call(f"Plot saved to ./vac_1_Cd_0.svg")
+        self.assertTrue(
[…]
