### 2022-09-20 f5b1493
Add tests for _save_plot() expected behaviour (renaming, printing info etc)
Former-commit-id: 9678c5178ba4f815cf272848fc42c2d11e38973d
