### 2022-09-20 7a9ea99
Update all plotting tests to match this behaviour
