### 2022-09-20 7fc471f
Add test to `test_input.py` for `distortion_metadata.json` renaming and info printing
