### 2022-09-20 03005d8
Update MPL plotting tests, rename and move directory to DATA_DIR
