+++ tests/test_shakenbreak.py
@@ -243,7 +243,7 @@ class ShakeNBreakTestCase(unittest.TestCase):  # integration testing ShakeNBreak
     # Now we test parsing of final energies and plotting
 
     @pytest.mark.mpl_image_compare(
-        baseline_dir="remote_baseline_plots",
+        baseline_dir="data/remote_baseline_plots",
         filename="vac_1_Cd_-2.png",
         style=f"{file_path}/../shakenbreak/shakenbreak.mplstyle",
         savefig_kwargs={"transparent": True, "bbox_inches": "tight"},
@@ -286,7 +286,7 @@ class ShakeNBreakTestCase(unittest.TestCase):  # integration testing ShakeNBreak
         return fig_dict["vac_1_Cd_-2"]
 
     @pytest.mark.mpl_image_compare(
-        baseline_dir="remote_baseline_plots",
+        baseline_dir="data/remote_baseline_plots",
         filename="vac_1_Cd_-1.png",
         style=f"{file_path}/../shakenbreak/shakenbreak.mplstyle",
         savefig_kwargs={"transparent": True, "bbox_inches": "tight"},
@@ -330,7 +330,7 @@ class ShakeNBreakTestCase(unittest.TestCase):  # integration testing ShakeNBreak
         return fig_dict["vac_1_Cd_-1"]
 
     @pytest.mark.mpl_image_compare(
-        baseline_dir="remote_baseline_plots",
+        baseline_dir="data/remote_baseline_plots",
         filename="vac_1_Cd_0.png",
         style=f"{file_path}/../shakenbreak/shakenbreak.mplstyle",
         savefig_kwargs={"transparent": True, "bbox_inches": "tight"},
