+++ tests/test_plotting.py
@@ -26,6 +26,7 @@ def if_present_rm(path):
 
 
 _file_path = os.path.dirname(__file__)
+_DATA_DIR = os.path.join(_file_path, "data")
 
 
 class PlottingDefectsTestCase(unittest.TestCase):
@@ -480,6 +481,7 @@ class PlottingDefectsTestCase(unittest.TestCase):
         # Saving to defect_dir subfolder in output_path
         fig, ax = plt.subplots(1, 1)
         defect_name = "vac_1_Cd_0"
+        if_present_rm(f"{os.path.join(self.VASP_CDTE_DATA_DIR, defect_name, defect_name)}.png")
         with patch("builtins.print") as mock_print, warnings.catch_warnings(
             record=True
         ) as w:
@@ -579,8 +581,8 @@ class PlottingDefectsTestCase(unittest.TestCase):
         if_present_rm(f"./{defect_name}_{current_datetime_minus1min}.svg")
 
     @pytest.mark.mpl_image_compare(
-        baseline_dir=f"{_file_path}/remote_baseline_plots",
-        filename="V$_{Cd}^{0}$_max_dist.png",
+        baseline_dir=f"{_DATA_DIR}/remote_baseline_plots",
+        filename="vac_1_Cd_0_max_dist.png",
         style=f"{_file_path}/../shakenbreak/shakenbreak.mplstyle",
         savefig_kwargs={"transparent": True, "bbox_inches": "tight"},
     )
@@ -596,8 +598,8 @@ class PlottingDefectsTestCase(unittest.TestCase):
         return fig
 
     @pytest.mark.mpl_image_compare(
-        baseline_dir=f"{_file_path}/remote_baseline_plots",
-        filename="V$_{Cd}^{0}$_fake_defect_name.png",
+        baseline_dir=f"{_DATA_DIR}/remote_baseline_plots",
+        filename="vac_1_Cd_0_fake_defect_name.png",
         style=f"{_file_path}/../shakenbreak/shakenbreak.mplstyle",
         savefig_kwargs={"transparent": True, "bbox_inches": "tight"},
     )
@@ -613,8 +615,8 @@ class PlottingDefectsTestCase(unittest.TestCase):
         return fig
 
     @pytest.mark.mpl_image_compare(
-        baseline_dir=f"{_file_path}/remote_baseline_plots",
-        filename="V$_{Cd}^{0}$_displacement.png",
+        baseline_dir=f"{_DATA_DIR}/remote_baseline_plots",
+        filename="vac_1_Cd_0_displacement.png",
         style=f"{_file_path}/../shakenbreak/shakenbreak.mplstyle",
         savefig_kwargs={"transparent": True, "bbox_inches": "tight"},
     )
@@ -631,8 +633,8 @@ class PlottingDefectsTestCase(unittest.TestCase):
         return fig
 
     @pytest.mark.mpl_image_compare(
-        baseline_dir=f"{_file_path}/remote_baseline_plots",
-        filename="V$_{Cd}^{0}$_maxdist_title_linecolor_label.png",
+        baseline_dir=f"{_DATA_DIR}/remote_baseline_plots",
+        filename="vac_1_Cd_0_maxdist_title_linecolor_label.png",
         style=f"{_file_path}/../shakenbreak/shakenbreak.mplstyle",
         savefig_kwargs={"transparent": True, "bbox_inches": "tight"},
     )
@@ -656,8 +658,8 @@ class PlottingDefectsTestCase(unittest.TestCase):
         return fig
 
     @pytest.mark.mpl_image_compare(
-        baseline_dir=f"{_file_path}/remote_baseline_plots",
-        filename="V$_{O}^{0}$_colors.png",
[…]
