### 2022-09-20 6a16f94
Update test_cli.py
Former-commit-id: 1e14276e820d6b42f267233fccbb66ba495865f9
