+++ tests/test_cli.py
@@ -89,8 +89,6 @@ class CLITestCase(unittest.TestCase):
                 or "vac_1_Cd_0" in i
             ):
                 shutil.rmtree(i)
-        if os.path.exists(f"{os.getcwd()}/distortion_plots"):
-            shutil.rmtree(f"{os.getcwd()}/distortion_plots")
 
         for defect in os.listdir(self.EXAMPLE_RESULTS):
             if os.path.isdir(f"{self.EXAMPLE_RESULTS}/{defect}"):
@@ -98,6 +96,11 @@ class CLITestCase(unittest.TestCase):
                     shutil.rmtree(f"{self.EXAMPLE_RESULTS}/{defect}/{dir}")
                     for dir in os.listdir(f"{self.EXAMPLE_RESULTS}/{defect}")
                     if "_from_" in dir
+                ]
+                [
+                    os.remove(f"{self.EXAMPLE_RESULTS}/{defect}/{file}")
+                    for file in os.listdir(f"{self.EXAMPLE_RESULTS}/{defect}")
+                    if file.endswith(".png") or file.endswith(".svg")
                 ]
             elif os.path.isfile(f"{self.EXAMPLE_RESULTS}/{defect}"):
                 os.remove(f"{self.EXAMPLE_RESULTS}/{defect}")
@@ -1436,17 +1439,18 @@ local_rattle: True"""
         # test warning when nothing parsed because defect folder not recognised
         os.chdir(self.EXAMPLE_RESULTS)
         with warnings.catch_warnings(record=True) as w:
-            result = runner.invoke(snb, ["parse", "-d", "defect"],
-                                   catch_exceptions=True)
+            result = runner.invoke(
+                snb, ["parse", "-d", "defect"], catch_exceptions=True
+            )
         self.assertTrue(any([warning.category == UserWarning for warning in w]))
         self.assertTrue(
             any(
                 [
                     str(warning.message)
                     == f"Energies could not be parsed for defect 'defect' in '.'. If these "
-                       f"directories are correct, check calculations have converged, and that "
-                       f"distortion subfolders match ShakeNBreak naming (e.g. "
-                       f"Bond_Distortion_xxx, Rattled, Unperturbed)"
+                    f"directories are correct, check calculations have converged, and that "
+                    f"distortion subfolders match ShakeNBreak naming (e.g. "
+                    f"Bond_Distortion_xxx, Rattled, Unperturbed)"
                     for warning in w
                 ]
             )
@@ -1684,7 +1688,9 @@ local_rattle: True"""
             f"{self.EXAMPLE_RESULTS}/{defect}",
             f"{self.EXAMPLE_RESULTS}/{defect}_defect_folder/{defect}",
         )
-        with open(f"{self.EXAMPLE_RESULTS}/{defect}_defect_folder/{defect}/{defect}.yaml", "w") as f:
+        with open(
+            f"{self.EXAMPLE_RESULTS}/{defect}_defect_folder/{defect}/{defect}.yaml", "w"
+        ) as f:
             f.write("")
         runner = CliRunner()
         result = runner.invoke(
@@ -1703,12 +1709,14 @@ local_rattle: True"""
             f"Saved results to {self.EXAMPLE_RESULTS}/{defect}_defect_folder/{defect}/{defect}.csv",
             result.output,
         )
-        with open(f"{self.EXAMPLE_RESULTS}/{defect}_defect_folder/{defect}/{defect}.csv") as f:
+        with open(
+            f"{self.EXAMPLE_RESULTS}/{defect}_defect_folder/{defect}/{defect}.csv"
+        ) as f:
             file = f.read()
         csv_content = (
[…]
