### 2022-09-20 0bbe6f3
Update test_local.py and test_shakenbreak.py and remove all instances of `distortion_plots` from files
Former-commit-id: 4c297efa2dc439e3f9c8d95d2689180bf988a2c1
