### 2022-09-20 b98f10f
Update notebooks to reflect new plot saving behaviour etc
