+++ tests/data/example_results/vac_1_Cd_-2/vac_1_Cd_-2.yaml
@@ -1,3 +1,3 @@
 distortions:
-  Rattled: -205.7
+  Rattled: -205.79
 Unperturbed: -205.8
