### 2022-09-20 cd8d15c
Add verbose option to all plotting commands
