### 2022-09-20 b6320ca
Update `pycodestyle` command
