### 2022-09-20 c577ff2
Fix plot tests
