### 2022-09-21 3bce8ca
Bump version number in `setup.py` 😅
