### 2022-09-21 0df77dc
Bump version number in `setup.py` 😅
Former-commit-id: 3bce8ca5a8b62f882e08dcf13d2fc1862dcc3c9a
