+++ setup.py
@@ -117,7 +117,7 @@ fonts = package_files("fonts/")
 
 setup(
     name="shakenbreak",
-    version="22.9.2",
+    version="22.9.21",
     description="Package to generate and analyse distorted defect structures, in order to "
     "identify ground-state and metastable defect configurations.",
     long_description="Python package to automatise the process of defect structure searching. "
