### 2022-09-21 9d29e40
Use `yaml` rather than `filecmp` to compare `yaml` files, to avoid issues with different formats but same info with tests on different systems
Former-commit-id: 2f4b3bd638e436be8910063613ae0c3af85b5558
