+++ tests/test_cli.py
@@ -7,8 +7,8 @@ import json
 import warnings
 from monty.serialization import loadfn
 import numpy as np
-import filecmp
 import subprocess
+import yaml
 
 # Pymatgen
 from pymatgen.core.structure import Structure
@@ -1505,12 +1505,11 @@ local_rattle: True"""
         self.assertTrue(
             os.path.exists(f"{self.DATA_DIR}/{code}/{defect}/{defect}.yaml")
         )
-        self.assertTrue(
-            filecmp.cmp(
-                f"{self.DATA_DIR}/{code}/{defect}/test_{defect}.yaml",
-                f"{self.DATA_DIR}/{code}/{defect}/{defect}.yaml",
-            )
-        )
+        with open(f"{self.DATA_DIR}/{code}/{defect}/test_{defect}.yaml", 'r') as test, \
+                open(f"{self.DATA_DIR}/{code}/{defect}/{defect}.yaml", 'r') as new:
+            test_yaml = yaml.safe_load(test)
+            new_yaml = yaml.safe_load(new)
+        self.assertDictEqual(test_yaml, new_yaml)
         os.remove(f"{self.DATA_DIR}/{code}/{defect}/{defect}.yaml")
 
         # CASTEP
@@ -1531,12 +1530,11 @@ local_rattle: True"""
         self.assertTrue(
             os.path.exists(f"{self.DATA_DIR}/{code}/{defect}/{defect}.yaml")
         )
-        self.assertTrue(
-            filecmp.cmp(
-                f"{self.DATA_DIR}/{code}/{defect}/test_{defect}.yaml",
-                f"{self.DATA_DIR}/{code}/{defect}/{defect}.yaml",
-            )
-        )
+        with open(f"{self.DATA_DIR}/{code}/{defect}/test_{defect}.yaml", 'r') as test, \
+                open(f"{self.DATA_DIR}/{code}/{defect}/{defect}.yaml", 'r') as new:
+            test_yaml = yaml.safe_load(test)
+            new_yaml = yaml.safe_load(new)
+        self.assertDictEqual(test_yaml, new_yaml)
         os.remove(f"{self.DATA_DIR}/{code}/{defect}/{defect}.yaml")
 
         # Espresso
@@ -1557,12 +1555,11 @@ local_rattle: True"""
         self.assertTrue(
             os.path.exists(f"{self.DATA_DIR}/{code}/{defect}/{defect}.yaml")
         )
-        self.assertTrue(
-            filecmp.cmp(
-                f"{self.DATA_DIR}/{code}/{defect}/test_{defect}.yaml",
-                f"{self.DATA_DIR}/{code}/{defect}/{defect}.yaml",
-            )
-        )
+        with open(f"{self.DATA_DIR}/{code}/{defect}/test_{defect}.yaml", 'r') as test, \
+                open(f"{self.DATA_DIR}/{code}/{defect}/{defect}.yaml", 'r') as new:
+            test_yaml = yaml.safe_load(test)
+            new_yaml = yaml.safe_load(new)
+        self.assertDictEqual(test_yaml, new_yaml)
         os.remove(f"{self.DATA_DIR}/{code}/{defect}/{defect}.yaml")
 
         # FHI-aims
@@ -1583,12 +1580,11 @@ local_rattle: True"""
         self.assertTrue(
             os.path.exists(f"{self.DATA_DIR}/{code}/{defect}/{defect}.yaml")
         )
-        self.assertTrue(
-            filecmp.cmp(
-                f"{self.DATA_DIR}/{code}/{defect}/test_{defect}.yaml",
-                f"{self.DATA_DIR}/{code}/{defect}/{defect}.yaml",
-            )
[…]
