### 2022-09-22 cfa7992
apply `isort` to sort/format imports and add isort to linting workflow
