### 2022-09-22 1241c8f
Don't include imported charge states in line plot, and make sure legend on top of all (including imported) datapoints
Former-commit-id: 80f8a50a318e887be3f6c55b8c033ee78c202d8b
