+++ shakenbreak/plotting.py
@@ -450,9 +450,11 @@ def _format_datapoints_from_other_chargestates(
     # Store indices of imported structures ("X%_from_Y") to plot differently later
     # comparison
     imported_indices = []
+    imported_energies = []
     for i, entry in enumerate(energies_dict["distortions"].keys()):
         if isinstance(entry, str) and "_from_" in entry:
             imported_indices.append(i)
+            imported_energies.append(energies_dict["distortions"][entry])
 
     # Reformat any "X%_from_Y" or "Rattled_from_Y" distortions to corresponding
     # (X) distortion factor or 0.0 for "Rattled"
@@ -507,6 +509,10 @@ def _format_datapoints_from_other_chargestates(
         sorted_distortions, sorted_energies = zip(
             *sorted(zip(keys, energies_dict["distortions"].values()))
         )
+        imported_indices = {  # unsorted_index: sorted_index
+            unsorted_index: sorted_energies.index(d)
+            for unsorted_index, d in zip(imported_indices, imported_energies)
+        }
         return imported_indices, keys, sorted_distortions, sorted_energies
     except ValueError:  # if keys and energies_dict["distortions"] are empty
         # (i.e. the only distortion is Rattled)
@@ -1286,10 +1292,18 @@ def plot_colorbar(
                 alpha=1,
             )
         else:
+            if imported_indices:  # Exclude datapoints from other charge states
+                non_imported_sorted_indices = [
+                    i for i in range(len(sorted_distortions)) if i not in imported_indices.values()
+                ]
+            else:
+                non_imported_sorted_indices = range(len(sorted_distortions))
+
+            # Plot non-imported distortions
             im = ax.scatter(  # Points for each distortion
-                sorted_distortions,
-                sorted_energies,
-                c=sorted_disp,
+                [sorted_distortions[i] for i in non_imported_sorted_indices],
+                [sorted_energies[i] for i in non_imported_sorted_indices],
+                c=[sorted_disp[i] for i in non_imported_sorted_indices],
                 ls="-",
                 s=50,
                 marker="o",
@@ -1298,8 +1312,8 @@ def plot_colorbar(
                 alpha=1,
             )
             ax.plot(  # Line connecting points
-                sorted_distortions,
-                sorted_energies,
+                [sorted_distortions[i] for i in non_imported_sorted_indices],
+                [sorted_energies[i] for i in non_imported_sorted_indices],
                 ls="-",
                 markersize=1,
                 marker="o",
@@ -1307,73 +1321,73 @@ def plot_colorbar(
                 label=legend_label,
             )
 
-        # Datapoints from other charge states
-        if imported_indices:
-            other_charges = len(
-                set(
-                    list(energies_dict["distortions"].keys())[i].split("_")[-1]
-                    for i in imported_indices.keys()
-                )
-            )  # number of other charge states whose distortions have been imported
-            for i, j in zip(imported_indices.keys(), range(other_charges)):
-                other_charge_state = int(
-                    list(energies_dict["distortions"].keys())[i].split("_")[-1]
-                )
-                sorted_i = imported_indices[i]  # index for the sorted dicts
-                ax.scatter(
-                    np.array(keys)[i],
-                    sorted_energies[sorted_i],
-                    c=sorted_disp[sorted_i],
-                    edgecolors="k",
-                    ls="-",
-                    s=50,
-                    marker=["s", "v", "<", ">", "^", "p", "X"][j],
-                    zorder=10,  # make sure it's on top of the other points
-                    cmap=colormap,
-                    norm=norm,
-                    alpha=1,
-                    label=f"From {'+' if other_charge_state > 0 else ''}{other_charge_state} "
-                    f"charge state",
-                )
+            # Datapoints from other charge states
[…]
