### 2022-09-22 80bb2a4
add pre-commit hook to catch formatting/linting issues at the commit stage (and allowing quick code linting with `pre-commit run --all-files`)
