### 2022-09-23 57724b7
update Contributing docs with pre-commit hook instructions
Former-commit-id: 4aa4113d824e860297eadd1833fa76c4595c9f91
