### 2022-09-23 4aa4113
update Contributing docs with pre-commit hook instructions
