### 2022-09-23 e3e42e8
add note/tip pannels to cli tutorial in the docs
Former-commit-id: bac978074a6a7853203e9562a4e0f71470dd09af
