### 2022-09-23 bac9780
add note/tip pannels to cli tutorial in the docs
