+++ docs/Generation.rst
@@ -21,34 +21,37 @@ we'll get a warning and we'll need to specify the defect site with the ``--defec
 
     $ snb-generate --bulk bulk_structure.cif --defect vac_1_Cd_POSCAR --min-charge -2 --max-charge 0 --defect-coords 0 0 0 --code VASP
 
-To specify additional distortion parameters, we can use a
-`config.yaml <https://github.com/SMTG-UCL/ShakeNBreak/blob/main/input_files/example_generate_config.yaml>`_
-file like the one below and use the ``--config`` flag to specify its path (i.e. ``snb-generate --config ./my_config.yaml``).
-A detailed description of all the parameters is available in the Python API section
-(:ref:`shakenbreak.input.Distortions class <api_input>`).
+.. NOTE::
+    To specify additional distortion parameters, we can use a
+    `config.yaml <https://github.com/SMTG-UCL/ShakeNBreak/blob/main/input_files/example_generate_config.yaml>`_
+    file like the one below and use the ``--config`` flag to specify its path (i.e. ``snb-generate --config ./my_config.yaml``).
+    A detailed description of all the parameters is available in the Python API section
+    (:ref:`shakenbreak.input.Distortions class <api_input>`).
 
-.. code:: yaml
+    .. code:: yaml
 
-    # Example config.yaml file for snb-generate
+        # Example config.yaml file for snb-generate
 
-    # General/Distortion section
-    oxidation_states:  # If not specified, the code will determine them
-       Cd: 2
-       Te: -2
-    distortion_increment: 0.1  # Bond distortion increment
-    distorted_elements:  # (Default = None, distorts nearest neighbours)
-        vac_Cd_1:
-            Cd  # Distort Cd atoms near the Cd interstitial
+        # General/Distortion section
+        oxidation_states:  # If not specified, the code will determine them
+        Cd: 2
+        Te: -2
+        distortion_increment: 0.1  # Bond distortion increment
+        distorted_elements:  # (Default = None, distorts nearest neighbours)
+            vac_Cd_1:
+                Cd  # Distort Cd atoms near the Cd interstitial
 
-    # Rattle section
-    stdev: 0.25  # Rattle standard deviation
-    d_min: 2.25  # Displacements that place atoms closer than d_min are penalised. (Default = 80% of auto-determined bulk bond length)
-    active_atoms: None  # Atoms to apply rattle displacement to. (Default = all atoms)
-    max_attempts: 5000  # Limit for how many attempted rattle moves are allowed a single atom; if this limit is reached an `Exception` is raised
-    max_disp: 2.0  # Rattle moves that yields a displacement larger than max_disp will always be rejected. Rarely occurs, mostly used as a safety net
-    local_rattle: False  # If True, rattle displacements will tail-off as we more away from the defect site. Not recommended as typically worsens performance.
+        # Rattle section
+        stdev: 0.25  # Rattle standard deviation
+        d_min: 2.25  # Displacements that place atoms closer than d_min are penalised. (Default = 80% of auto-determined bulk bond length)
+        active_atoms: None  # Atoms to apply rattle displacement to. (Default = all atoms)
+        max_attempts: 5000  # Limit for how many attempted rattle moves are allowed a single atom; if this limit is reached an `Exception` is raised
+        max_disp: 2.0  # Rattle moves that yields a displacement larger than max_disp will always be rejected. Rarely occurs, mostly used as a safety net
+        local_rattle: False  # If True, rattle displacements will tail-off as we more away from the defect site. Not recommended as typically worsens performance.
 
-To display additional information about the generated distortions we can set the ``--verbose`` flag.
+
+.. TIP::
+    To display additional information about the generated distortions we can set the ``--verbose`` flag.
 
 For many defects
[…]
