### 2022-09-23 f67d5ef
refactor `_calc_number_electrons`, `_get_bulk_comp`  and _apply_rattle_bond_distortions to use Defect() object rather than dict
Former-commit-id: b357935e111bf8becd76d110c07fc60938faf840
