+++ shakenbreak/input.py
@@ -1,3 +1,6 @@
+# TODO: Adding a "# refactored" decorator to methods/functions that
+# have been refactored to the Defect class
+
 """
 Module containing functions to generate rattled and bond-distorted structures,
 as well as input files to run Gamma point relaxations with `VASP`, `CP2K`,
@@ -17,6 +20,7 @@ from ase.calculators.aims import Aims
 from ase.calculators.castep import Castep
 from ase.calculators.espresso import Espresso
 from monty.serialization import loadfn
+from pymatgen.analysis.defects.core import Defect
 from pymatgen.core.structure import Composition, Element, Structure
 from pymatgen.io.ase import AseAtomsAdaptor
 from pymatgen.io.cp2k.inputs import Cp2kInput
@@ -202,6 +206,7 @@ def _create_vasp_input(
             potcar_settings
         )  # files empties `potcar_settings dict` (via pop()), so make a
         # deepcopy each time
+        # TODO: refactor vasp.write_vasp_gam_files
         vasp.write_vasp_gam_files(
             single_defect_dict=single_defect_dict,
             input_dir=f"{output_path}/{defect_name}/{distortion}",
@@ -210,36 +215,22 @@ def _create_vasp_input(
         )
 
 
-def _get_bulk_comp(defect_dict) -> Composition:
+# refactored
+def _get_bulk_comp(defect_object) -> Composition:
     """
     Convenience function to determine the chemical composition of the bulk
     structure for a given defect. Useful for auto-determing oxidation states.
 
     Args:
-        defect_dict (:obj:`dict`):
-            Defect dictionary in the
-            `doped.pycdt.core.defectsmaker.ChargedDefectsStructures()` format.
+        defect_object (:obj:`Defect`):
+           pymatgen.analysis.defects.core.Defect() object.
 
     Returns:
         Pymatgen Composition object for the bulk structure of the defect.
     """
-    defect_struc = defect_dict["supercell"]["structure"].copy()
-    defect_sites = defect_struc.sites
-    defect_type = defect_dict["defect_type"]
+    bulk_structure = defect_object.structure
 
-    if defect_type == "vacancy":
-        bulk_sites = defect_sites + [defect_dict["unique_site"]]
-
-    elif defect_type == "interstitial":
-        bulk_sites = defect_sites
-        bulk_sites.remove(defect_dict["unique_site"])
-
-    elif defect_type in ["substitution", "antisite"]:
-        bulk_sites = defect_sites
-        bulk_sites.remove(defect_dict["bulk_supercell_site"])
-        bulk_sites += [defect_dict["unique_site"]]
-
-    return Structure.from_sites(bulk_sites).composition
+    return bulk_structure.composition
 
 
 def _most_common_oxi(element) -> int:
@@ -264,19 +255,19 @@ def _most_common_oxi(element) -> int:
     return most_common.oxi_state
 
 
+# refactored
 def _calc_number_electrons(
-    defect_dict: dict,
+    defect_object: Defect,
     oxidation_states: dict,
     verbose: bool = False,
 ) -> int:
     """
     Calculates the number of extra/missing electrons of the defect species
-    (in `defect_dict`) based on `oxidation_states`.
+    (in `defect_object`) based on `oxidation_states`.
 
     Args:
-        defect_dict (:obj:`dict`):
-            Defect dictionary in the
-            `doped.pycdt.core.defectsmaker.ChargedDefectsStructures()` format.
+        defect_object (:obj:`Defect`):
+            pymatgen.analysis.defects.core.Defect object.
         oxidation_states (:obj:`dict`):
             Dictionary with oxidation states of the atoms in the material (e.g.
             {"Cd": +2, "Te": -2}).
[…]
