### 2022-09-23 ca7a7dc
Add PyPI downloads per month badge to readme and docs, currently ~400/month!
