+++ docs/index.rst
@@ -17,6 +17,9 @@
 
 .. image:: https://joss.theoj.org/papers/6545bcc1a0439b16360ace684ac5aa25/status.svg
  :target: https://joss.theoj.org/papers/6545bcc1a0439b16360ace684ac5aa25
+
+.. image:: https://img.shields.io/pypi/dm/shakenbreak
+ :target: https://shakenbreak.readthedocs.io/en/latest/
 |
 
 Welcome to ShakeNBreak!
