### 2022-09-23 c212eee
Add PyPI downloads per month badge to readme and docs, currently ~400/month!
Former-commit-id: ca7a7dcd6d45e3f0bd210677ef372b3aabb8e290
