### 2022-09-23 9744910
(Tiny) update to structure comment format (two underscores for both spacings)
