### 2022-09-23 1340649
Fix so that "Rattled_from_-1" etc is correctly parsed by format_defect_names() (previously unrecognised) and add tests for this function
