### 2022-09-23 ecaabd4
Fix so that "Rattled_from_-1" etc is correctly parsed by format_defect_names() (previously unrecognised) and add tests for this function
Former-commit-id: 1340649bf0d414764e84c9afb53d25a0d804c905
