### 2022-09-23 9ade833
Update format_distortion_directory_name() to give "Rattled_from_2" instead of "Bond_Distortion_Rattled_from_2" etc, and add tests for format_distortion_directory_name()
Former-commit-id: cc251215c2aeac032e89e2d9d7826eb1873a12ff
