### 2022-09-23 6434a5b
Update `energy_lowering_distortions` to be able to handle "Rattled_from_2" etc with `get_energy_lowering_distortions()` and add test for this
