### 2022-09-23 d6d01c8
Update `energy_lowering_distortions` to be able to handle "Rattled_from_2" etc when pruning across charge states and add test for this
Former-commit-id: 73211296bf5812bdb74cd50e6d1b5985c78b67fe
