### 2022-09-23 7321129
Update `energy_lowering_distortions` to be able to handle "Rattled_from_2" etc when pruning across charge states and add test for this
