### 2022-09-23 2927b9a
Update `test_cli.py` and `test_shakenbreak.py` to match other updates
