### 2022-09-23 3ec4972
Update `test_cli.py` and `test_shakenbreak.py` to match other updates
Former-commit-id: 2927b9a24eed7faf32f9221ff1a703175da89dbe
