### 2022-09-24 c5bc06d
Add error for `write_groundstate_structure()` if no defect directories detected (previously would run with no output but doing nothing), and add test for this
Former-commit-id: 78744e2c190f4957de8cb930680cbec75d1e0c62
