+++ tests/test_cli.py
@@ -9,9 +9,11 @@ import warnings
 
 import numpy as np
 import yaml
+
 # Click
 from click.testing import CliRunner
 from monty.serialization import loadfn
+
 # Pymatgen
 from pymatgen.core.structure import Structure
 from pymatgen.io.vasp.inputs import Incar, Poscar
@@ -1504,8 +1506,9 @@ local_rattle: True"""
         self.assertTrue(
             os.path.exists(f"{self.DATA_DIR}/{code}/{defect}/{defect}.yaml")
         )
-        with open(f"{self.DATA_DIR}/{code}/{defect}/test_{defect}.yaml", 'r') as test, \
-                open(f"{self.DATA_DIR}/{code}/{defect}/{defect}.yaml", 'r') as new:
+        with open(
+            f"{self.DATA_DIR}/{code}/{defect}/test_{defect}.yaml", "r"
+        ) as test, open(f"{self.DATA_DIR}/{code}/{defect}/{defect}.yaml", "r") as new:
             test_yaml = yaml.safe_load(test)
             new_yaml = yaml.safe_load(new)
         self.assertDictEqual(test_yaml, new_yaml)
@@ -1529,8 +1532,9 @@ local_rattle: True"""
         self.assertTrue(
             os.path.exists(f"{self.DATA_DIR}/{code}/{defect}/{defect}.yaml")
         )
-        with open(f"{self.DATA_DIR}/{code}/{defect}/test_{defect}.yaml", 'r') as test, \
-                open(f"{self.DATA_DIR}/{code}/{defect}/{defect}.yaml", 'r') as new:
+        with open(
+            f"{self.DATA_DIR}/{code}/{defect}/test_{defect}.yaml", "r"
+        ) as test, open(f"{self.DATA_DIR}/{code}/{defect}/{defect}.yaml", "r") as new:
             test_yaml = yaml.safe_load(test)
             new_yaml = yaml.safe_load(new)
         self.assertDictEqual(test_yaml, new_yaml)
@@ -1554,8 +1558,9 @@ local_rattle: True"""
         self.assertTrue(
             os.path.exists(f"{self.DATA_DIR}/{code}/{defect}/{defect}.yaml")
         )
-        with open(f"{self.DATA_DIR}/{code}/{defect}/test_{defect}.yaml", 'r') as test, \
-                open(f"{self.DATA_DIR}/{code}/{defect}/{defect}.yaml", 'r') as new:
+        with open(
+            f"{self.DATA_DIR}/{code}/{defect}/test_{defect}.yaml", "r"
+        ) as test, open(f"{self.DATA_DIR}/{code}/{defect}/{defect}.yaml", "r") as new:
             test_yaml = yaml.safe_load(test)
             new_yaml = yaml.safe_load(new)
         self.assertDictEqual(test_yaml, new_yaml)
@@ -1579,8 +1584,9 @@ local_rattle: True"""
         self.assertTrue(
             os.path.exists(f"{self.DATA_DIR}/{code}/{defect}/{defect}.yaml")
         )
-        with open(f"{self.DATA_DIR}/{code}/{defect}/test_{defect}.yaml", 'r') as test, \
-                open(f"{self.DATA_DIR}/{code}/{defect}/{defect}.yaml", 'r') as new:
+        with open(
+            f"{self.DATA_DIR}/{code}/{defect}/test_{defect}.yaml", "r"
+        ) as test, open(f"{self.DATA_DIR}/{code}/{defect}/{defect}.yaml", "r") as new:
             test_yaml = yaml.safe_load(test)
             new_yaml = yaml.safe_load(new)
         self.assertDictEqual(test_yaml, new_yaml)
[…]
