### 2022-09-25 db4b597
refactor `apply_snb_distortions` to take defect object rather than defect dict
Former-commit-id: 251f2f8abed2764b9b7edd7fd27704226ae296f1
