+++ shakenbreak/input.py
@@ -233,6 +233,36 @@ def _get_bulk_comp(defect_object) -> Composition:
     return bulk_structure.composition
 
 
+def _get_bulk_defect_site(defect_object):
+    """Get defect site in the bulk structure (e.g.
+    for a P substitution on Si, get the original Si site).
+    """
+    defect_type = defect_object.as_dict()["@class"].lower()
+    if defect_type in ["antisite", "substitution"]:
+        # get bulk_site
+        poss_deflist = sorted(
+            defect_object.structure.get_sites_in_sphere(  # Defect().structure is bulk_structure
+                defect_object.site.coords, 0.01, include_index=True
+            ),
+            key=lambda x: x[1],
+        )
+        if not poss_deflist:
+            raise ValueError(
+                "Error in defect object generation; could not find substitution "
+                f"site inside bulk structure for {defect_object.name}"
+            )
+        defindex = poss_deflist[0][2]
+        sub_site_in_bulk = defect_object.structure[
+            defindex
+        ]  # bulk site of substitution
+
+        unique_site = sub_site_in_bulk
+
+    else:
+        unique_site = defect_object.site
+    return unique_site
+
+
 def _most_common_oxi(element) -> int:
     """
     Convenience function to get the most common oxidation state of an element, using pymatgen's
@@ -300,22 +330,10 @@ def _calc_number_electrons(
 
     elif defect_type in ["antisite", "substitution"]:
         # get bulk_site
-        poss_deflist = sorted(
-            defect_object.structure.get_sites_in_sphere(  # Defect().structure is bulk_structure
-                defect_object.site.coords, 0.01, include_index=True
-            ),
-            key=lambda x: x[1],
-        )
-        if not poss_deflist:
-            raise ValueError(
-                "Error in defect object generation; could not find substitution "
-                f"site inside bulk structure for {defect_name}"
-            )
-        defindex = poss_deflist[0][2]
-        sub_site_in_bulk = defect_object.structure[
-            defindex
-        ]  # bulk site of substitution
-        site_specie = sub_site_in_bulk.specie.symbol  # Species occuping *bulk* site
+        sub_site_in_bulk = _get_bulk_defect_site(
+            defect_object
+        )  # bulk site of substitution
+        site_specie = sub_site_in_bulk.specie.symbol  # Species occuping the *bulk* site
         substituting_specie = str(
             defect_object.site.specie.symbol
         )  # Current species occupying the defect site (e.g. the substitution)
@@ -363,8 +381,9 @@ def _calc_number_neighbours(num_electrons: int) -> int:
 
 # Main functions
 
+# refactored
 # Input arguments have beeen refactored to take Defect() object
-# No we need to refactor the output dictionary (and for that the distort() function)
+# No we need to refactor the output dictionary
 def _apply_rattle_bond_distortions(
     defect_object: Defect,
     num_nearest_neighbours: int,
@@ -384,7 +403,7 @@ def _apply_rattle_bond_distortions(
             - defect site index (other defect types).
 
     Args:
-        defect_object (:obj:`dict`):
+        defect_object (:obj:`Defect`):
             pymatgen.analysis.defects.core.Defect()
         num_nearest_neighbours (:obj:`int`):
             Number of defect nearest neighbours to apply bond distortions to.
@@ -449,6 +468,7 @@ def _apply_rattle_bond_distortions(
     defect_type = str(defect_object.as_dict()["@class"].lower())
     bulk_supercell_site = defect_object.site
     defect_structure = defect_object.defect_structure
+
     if defect_type == "vacancy":  # for vacancies, we need to use fractional coordinates
         # (no atom site in structure!)
[…]
