### 2022-09-25 5e1c151
update __init__ method of Distortions() class
Former-commit-id: 23969a65af5a12ad9d84e7a75f826f349d4aa29f
