+++ shakenbreak/input.py
@@ -237,7 +237,7 @@ def _get_bulk_defect_site(defect_object):
     """Get defect site in the bulk structure (e.g.
     for a P substitution on Si, get the original Si site).
     """
-    defect_type = defect_object.as_dict()["@class"].lower()
+    defect_type = str(defect_object.as_dict()["@class"].lower())
     if defect_type in ["antisite", "substitution"]:
         # get bulk_site
         poss_deflist = sorted(
@@ -761,12 +761,15 @@ def apply_snb_distortions(
 
 # TODO: As discussed, need to refactor Distortions() to take a dictionary of Defect() objects
 # rather than a dict of (doped/pmg/pycdt) defect_dict
+# We need a way to store the defect charge states defined by the user though, as with
+# the get_charge_states() method we need to give the padding
+# a)
 
 
 class Distortions:
     """
     Class to apply rattle and bond distortion to all defects in `defects_dict`
-    (in `doped` `ChargedDefectsStructures()` format).
+    (each defect as a pymatgen.analysis.defects.core.Defect() object).
     """
 
     def __init__(
@@ -784,7 +787,12 @@ class Distortions:
         """
         Args:
             defects_dict (:obj:`dict`):
-                Dictionary of pymatgen-analysis-defects.core.Defect() objects
+                Dictionary of pymatgen.analysis.defects.core.Defect() objects.
+                E.g.: {
+                    "vacancies": [Vacancy(), ...],
+                    "interstitials": [Interstitial(), ...],
+                    "substitutions": [Substitution(), ...],
+                }
             oxidation_states (:obj:`dict`):
                 Dictionary of oxidation states for species in your material,
                 used to determine the number of defect neighbours to distort
@@ -858,34 +866,39 @@ class Distortions:
 
         # check if all expected oxidation states are provided
         if "bulk" in self.defects_dict:
-            bulk_comp = self.defects_dict["bulk"]["supercell"]["structure"].composition
+            bulk_comp = self.defects_dict[
+                "bulk"
+            ].structure.composition  # TODO: check this after doped is updated
             guessed_oxidation_states = bulk_comp.oxi_state_guesses()[0]
 
-        else:  # determine bulk composition from first defect in dict
-            defect_subdict = list(self.defects_dict.values())[0][0]
-            bulk_comp = _get_bulk_comp(defect_subdict)
+        else:  # determine bulk composition from first Defect in dict
+            defect_object = list(self.defects_dict.values())[0][
+                0
+            ]  # TODO: check this after doped is updated
+            bulk_comp = _get_bulk_comp(defect_object)
             guessed_oxidation_states = bulk_comp.oxi_state_guesses()[0]
 
         if "substitutions" in self.defects_dict:
             for substitution in self.defects_dict["substitutions"]:
+                defect_type = str(substitution.as_dict()["@class"].lower())
                 if (
-                    substitution["defect_type"] == "substitution"
+                    defect_type == "substitution"
                     and substitution["bulk_supercell_site"].specie.symbol
                     not in guessed_oxidation_states
                 ):
                     # extrinsic substituting species not in bulk composition
-                    extrinsic_specie = substitution["bulk_supercell_site"].specie.symbol
+                    extrinsic_specie = substitution.defect_site.specie.symbol
                     likely_substitution_oxi = _most_common_oxi(extrinsic_specie)
                     guessed_oxidation_states[extrinsic_specie] = likely_substitution_oxi
 
         if "interstitials" in self.defects_dict:
             for interstitial in self.defects_dict["interstitials"]:
                 if (
-                    interstitial["bulk_supercell_site"].specie.symbol
+                    interstitial.defect_site.specie.symbol
                     not in guessed_oxidation_states
                 ):
[…]
