### 2022-09-26 #4697 @VHchavez
Hello @rkurchin & @ireaml, ShakeNBreak looks fantastic.
I, unfortunately, don't have much experience with solids, nor have I run any calculations using the codes mentioned (VASP, CP2K, Quantum-Espresso, CASTEP & FHI-aims). Thus I don't think I'm fit for this review.
