### 2022-09-29 #4697 @rkurchin
👋 @mkhorton and @obaica, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
