### 2022-09-29 #4697 @obaica
I am more than willing, and it is my honor to review this submission for JOSS.

What is the exact process we need to do next?

Thank you very much.

Gang Tang
________________________________
Advanced Research Institute of Multidisciplinary Science
Beijing Institute of Technology
No. 5 Zhongguancun South Street, Haidian District
Beijing, 100081 P. R. China.
***@***.*** or ***@***.***
Researchgate Homepage:
https://www.researchgate.net/profile/Gang_Tang<https://www.researchgate.net/profile/Gang_Tang3>3
________________________________
发件人: Rachel Kurchin ***@***.***>
发送时间: 2022年9月29日 22:20
收件人: openjournals/joss-reviews ***@***.***>
抄送: Gang Tang ***@***.***>; Mention ***@***.***>
主题: Re: [openjournals/joss-reviews] [PRE REVIEW]: ShakeNBreak: Navigating the defect configurational landscape (Issue #4697)

👋 @mkhorton<https://github.com/mkhorton> and @obaica<https://github.com/obaica>, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
