### 2022-09-29 #4697 @rkurchin
Great, thanks so much! I'll add you as a reviewer, and once we have at least two reviewers, a new issue will be opened where there will be a checklist to guide you through the process. If you want to learn more now, you can read more details [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) -> note particularly the criteria and checklist links on the navigation bar to the left.
