### 2022-09-29 #4697 @rkurchin
Great! We don't normally impose strict deadlines since the time commitment can vary a fair bit depending on how elaborate the software is and how much feedback you have as a reviewer, so I'm generally more interested that you can get started relatively soon (within a week or so) than exactly how long it will take. I think unless it's an unusually large/intricate piece of software, it would be good if you could get your first set of feedback (if any) turned around within a month or so, ballpark.
