### 2022-10-03 ab4cbb8
If positive energies encountered for ionic steps after 5th ionic step, ignore and rename. Add tests for this
Former-commit-id: 6d36aa23d87564fbd6b82168d8f5724a8b18409f
