### 2022-10-03 6d36aa2
If positive energies encountered for ionic steps after 5th ionic step, ignore and rename. Add tests for this
