### 2022-10-03 3873351
Ignore "*High_Energy*" folders with io.parse_energies() and test this
Former-commit-id: df44452f81c74f956b74828878dc21132aa08490
