### 2022-10-03 df44452
Ignore "*High_Energy*" folders with io.parse_energies() and test this
