### 2022-10-03 2de0566
Add test in test_shakenbreak.py for this behaviour
Former-commit-id: ae9604d30995d165b034da3a1a508604403771df
