### 2022-10-03 4860990
Ignore "*High_Energy*" with `get_structures()` and add test for this
Former-commit-id: e266dc7439803960a81e4d47bea077513f755c5d
