### 2022-10-03 e266dc7
Ignore "*High_Energy*" with `get_structures()` and add test for this
