### 2022-10-03 54e85a3
Add tests to `test_groundstate` and `test_regenerate` that "*High_Energy*" folders ignored as expected
