### 2022-10-03 4999464
Update SnB_run.sh to do same "High_Energy" renaming if forces errors (i.e. EDDDAV or ZPOTRF) are encountered, and add tests for this
Former-commit-id: 8a7d9681462d5aa99354ad2f2c305600cc53d181
