+++ tests/test_cli.py
@@ -114,7 +114,7 @@ class CLITestCase(unittest.TestCase):
         folder = "Bond_Distortion_-60.0%_from_0"
         for charge in [-1, -2]:
             if os.path.exists(
-                    os.path.join(self.EXAMPLE_RESULTS, f"vac_1_Cd_{charge}", folder)
+                os.path.join(self.EXAMPLE_RESULTS, f"vac_1_Cd_{charge}", folder)
             ):
                 shutil.rmtree(
                     os.path.join(self.EXAMPLE_RESULTS, f"vac_1_Cd_{charge}", folder)
@@ -122,13 +122,16 @@ class CLITestCase(unittest.TestCase):
         folder = "Bond_Distortion_20.0%_from_-1"
         for charge in [0, -2]:
             if os.path.exists(
-                    os.path.join(self.EXAMPLE_RESULTS, f"vac_1_Cd_{charge}", folder)
+                os.path.join(self.EXAMPLE_RESULTS, f"vac_1_Cd_{charge}", folder)
             ):
                 shutil.rmtree(
                     os.path.join(self.EXAMPLE_RESULTS, f"vac_1_Cd_{charge}", folder)
                 )
-        if_present_rm(os.path.join(self.EXAMPLE_RESULTS,
-                                     "vac_1_Cd_0/Bond_Distortion_-48.0%_High_Energy"))
+        if_present_rm(
+            os.path.join(
+                self.EXAMPLE_RESULTS, "vac_1_Cd_0/Bond_Distortion_-48.0%_High_Energy"
+            )
+        )
 
     def test_snb_generate(self):
         runner = CliRunner()
@@ -1271,7 +1274,6 @@ local_rattle: True"""
             ["snb-run", "-v", "-a"], stdout=subprocess.PIPE, stderr=subprocess.PIPE
         )
         out = str(proc.communicate()[0])
-        print(out)
         self.assertIn(
             "No defect folders (with names ending in a number (charge state)) found in "
             "current directory",
@@ -1309,13 +1311,12 @@ energy  without entropy=        7.99185422  energy(sigma->0) =        7.99185422
         self.assertNotIn("Running job for Bond_Distortion_10.0%", out)
         self.assertNotIn("this vac_1_Ti_0_10.0% job_file", out)  # job submit command
         self.assertIn(
-            "Positive energies encountered for Bond_Distortion_10.0%, ignoring and "
+            "Positive energies or forces error encountered for Bond_Distortion_10.0%, ignoring and "
             "renaming to Bond_Distortion_10.0%_High_Energy",
             out,
         )
         self.assertFalse(os.path.exists("Bond_Distortion_10.0%"))
         self.assertTrue(os.path.exists("Bond_Distortion_10.0%_High_Energy"))
-        if_present_rm("job_file")
         if_present_rm("Bond_Distortion_10.0%_High_Energy/job_file")
         if_present_rm("Bond_Distortion_10.0%_High_Energy/OUTCAR")
 
@@ -1332,7 +1333,7 @@ energy  without entropy=        7.99185422  energy(sigma->0) =        7.99185422
         self.assertNotIn("Running job for Bond_Distortion_10.0%", out)
         self.assertNotIn("this vac_1_Ti_0_10.0% job_file", out)  # job submit command
         self.assertNotIn(
-            "Positive energies encountered for Bond_Distortion_10.0%, ignoring and "
+            "Positive energies or forces encountered for Bond_Distortion_10.0%, ignoring and "
             "renaming to Bond_Distortion_10.0%_High_Energy",
             out,
         )
@@ -1342,9 +1343,7 @@ energy  without entropy=        7.99185422  energy(sigma->0) =        7.99185422
[…]
