### 2022-10-04 578ac7f
Update warning in `energy_lowering_distortions` about consistent high energies for distortions to mention likelihood of unreasonable charge state(s), and update test
