### 2022-10-04 ddf51f2
Update docs as discussed, with info about forces errors handling and unreasonable defect charge states
