+++ docs/Tips.rst
@@ -5,19 +5,30 @@ Tricky Relaxations
 -------------------
 
 If certain relaxations are not converging after multiple continuation calculations (i.e. if :code:`snb-run` keeps
-resubmitting certain relaxations), this is likely due to an error in the underlying calculation, and so you should
-check the calculation output files to see if this requires fixing. Often this may require changing a specific input file
-setting (e.g. in the :code:`INCAR` for :code:`VASP`), and copying the updated input files to other directories for which
-relaxations are struggling to converge.
+resubmitting certain relaxations), this is likely due to an error in the underlying calculation and/or unreasonable
+interatomic distances causing high / positive energies and extreme forces.
 
 - For :code:`VASP`, a common culprit is :code:`EDWAV` in the output file, which can typically be avoided by reducing
-  :code:`NCORE` and/or :code:`KPAR`. Another is :code:`ZPOTRF` which can happen in rare cases and suggests too large
-  initial forces, so typically easiest to just ignore this specific distortion.
+  :code:`NCORE` and/or :code:`KPAR`. Other errors related to forces / unreasonable interatomic distances (like
+  :code:`EDDDAV` or :code:`ZPOTRF`), as well as cases with positive energies, are automatically handled by
+  :code:`snb-run` (folder is renamed to :code:`Bond_Distortion_X_High_Energy` and subsequently ignored).
+    - If some relaxations are still not converging after multiple continuations, you should check the calculation output
+      files to see if this requires fixing. Often this may require changing a specific input file setting (e.g. in the
+      :code:`INCAR` for :code:`VASP`), and copying the updated input files to other directories for which relaxations are
+      struggling to converge.
+- For codes other than :code:`VASP`, forces errors or high / positive energies are not automatically handled, and so in
+  the rare cases where this occurs, you should rename the folder(s) to :code:`Bond_Distortion_X_High_Energy` and
+  :code:`ShakeNBreak` will subsequently ignore them.
 
 If the calculation outputs show that the relaxation is proceeding fine, without any errors, just not converging to
 completion, then other input settings such as the ionic relaxation algorithm (:code:`IBRION` in :code:`VASP`),
 electronic minimisation algorithm (:code:`ALGO` in :code:`VASP`) or real space force projection (:code:`LREAL`
 in :code:`VASP`) should be adjusted to aid convergence.
+
+In the other rare case where all distortions yield high energies, relative to the :code:`Unperturbed` structure, this is
+typically indicative of an unreasonable defect charge state (with the extreme excess charge inducing many false local
+minima on the PES). :code:`ShakeNBreak` will print a warning in these cases, with advice on how to proceed if this is
+not the case and the charge state is reasonable.
 
 
 Hard/Ionic Materials
@@ -27,31 +38,35 @@ The default bond distortion range of -60% to +60%, and the default rattling stan
 extreme in the case of hard/ionic/oxide materials which typically yield larger forces in response to bond distortion.
 If this is the case for your material, it will manifest in the form of:
 
-- High energies / unconverging calculations for the ±60% endpoints. Here you should adjust the distortion range to
-  exclude these points (e.g. :code:`bond_distortions = np.arange(-0.5, 0.501, 0.1)`), or just ignore these calculations.
+- High energies / non-converging calculations for the ±60% endpoints. As mentioned in :ref:`Tricky Relaxations` above,
+  these are automatically handled by :code:`snb-run` for :code:`VASP`, but for other codes you should rename the
+  folder(s) to :code:`Bond_Distortion_X_High_Energy` and :code:`ShakeNBreak` will subsequently ignore them.
+.. Here you should adjust the distortion range to exclude these points (e.g. :code:`bond_distortions = np.arange(-0.5, 0.501, 0.1)`), or just ignore these calculations.
 
[…]
