### 2022-10-04 de12e79
Update distortion generation as discussed, with skipping and warning for interatomic distances less than 1 Angstrom (unless H about), and add tests for this
Former-commit-id: 3c08f58556e9c9c0b53f120d2adedb4aa951fb59
