+++ tests/test_input.py
@@ -5,14 +5,15 @@ import os
 import pickle
 import shutil
 import unittest
+import warnings
 from unittest.mock import patch
 
 import numpy as np
 from ase.calculators.aims import Aims
-from pymatgen.core.structure import Composition, Element, PeriodicSite, Structure
+from pymatgen.core.structure import Composition, PeriodicSite, Structure
 from pymatgen.io.vasp.inputs import Poscar
 
-from shakenbreak import distortions, input, io, vasp
+from shakenbreak import distortions, input, vasp
 
 
 def if_present_rm(path):
@@ -1486,8 +1487,7 @@ class InputTestCase(unittest.TestCase):
         # The input_file option is tested through the test for `generate_all()`
         # (in `test_cli.py`)
 
-    @patch("builtins.print")
-    def test_apply_distortions(self, mock_print):
+    def test_apply_distortions(self):
         """Test method apply_distortions"""
         # check files are not written if `apply_distortions()` method is used
         for i in self.cdte_defect_folders:
@@ -1505,6 +1505,126 @@ class InputTestCase(unittest.TestCase):
             verbose=False,
         )
         self.assertFalse(os.path.exists("vac_1_Cd_0"))
+
+        # test bond distortions with interatomic distances less than 1 Angstrom are omitted,
+        # unless hydrogen involved
+        with warnings.catch_warnings(record=True) as w:
+            bond_distortions = list(np.arange(-1.0, 0.01, 0.05))
+            dist = input.Distortions(
+                {"vacancies": [reduced_V_Cd_dict]},
+                bond_distortions=bond_distortions,
+            )
+            distortion_defect_dict, distortion_metadata = dist.apply_distortions(
+                verbose=True
+            )
+        self.assertFalse(os.path.exists("vac_1_Cd_0"))
+
+        self.assertEqual(
+            len([warning for warning in w if warning.category == UserWarning]), 5
+        )
+        message_1 = (
+            "Bond_Distortion_-100.0% for defect vac_1_Cd gives an interatomic "
+            "distance less than 1.0 Å (0.0 Å), which is likely to give explosive "
+            "forces. Omitting this distortion."
+        )
+        message_2 = (
+            "Bond_Distortion_-95.0% for defect vac_1_Cd gives an interatomic "
+            "distance less than 1.0 Å (0.23 Å), which is likely to give explosive "
+            "forces. Omitting this distortion."
+        )
+        message_3 = (
+            "Bond_Distortion_-90.0% for defect vac_1_Cd gives an interatomic "
+            "distance less than 1.0 Å (0.46 Å), which is likely to give explosive "
+            "forces. Omitting this distortion."
+        )
+        message_4 = (
+            "Bond_Distortion_-85.0% for defect vac_1_Cd gives an interatomic "
+            "distance less than 1.0 Å (0.69 Å), which is likely to give explosive "
+            "forces. Omitting this distortion."
+        )
+        message_5 = (
+            "Bond_Distortion_-80.0% for defect vac_1_Cd gives an interatomic "
+            "distance less than 1.0 Å (0.93 Å), which is likely to give explosive "
+            "forces. Omitting this distortion."
+        )
+        self.assertTrue(
+            all(
+                [
+                    any([message == str(warning.message) for warning in w])
+                    for message in [
+                        message_1,
+                        message_2,
+                        message_3,
+                        message_4,
+                        message_5,
+                    ]
+                ]
+            )
+        )
+        V_Cd_distortions_dict = distortion_defect_dict["vac_1_Cd"]["charges"][0]["structures"][
+            "distortions"]
+        self.assertEqual(len(V_Cd_distortions_dict), 16)
[…]
