### 2022-10-04 3678f49
Make `ref_struct` docstrings & help messages a bit more informative
