### 2022-10-04 39b220f
Update docs to show CLI commands running with default minimal arguments first (to show ease-of-use and user-friendliness etc) then with optional flags for greater control
Former-commit-id: 5a32ba5919c6835fca18a354625ee176ab7366de
