+++ docs/Analysis.rst
@@ -6,21 +6,29 @@ Analysis & plotting
 Parsing
 ----------
 
-To parse the final energies of the geometry optimisations for a specific defect:
+To parse the final energies of the geometry optimisations for a specific defect, we can simply run the command
+:code:`snb-parse` within the defect folder:
 
 .. code:: bash
 
-    $ snb-parse --code vasp --defect vac_1_Cd_0 --path defects_folder
+    $ snb-parse
 
-Where ``defects_folder`` is the path to the top level directory containing the defect folder,
-and is only required if different from the current directory.
+Alternatively, we can run from a different directory and specify the defect to parse (as well as other options, see
+``snb-parse -h``):
+
+.. code:: bash
+
+    $ snb-parse --defect vac_1_Cd_0 --path defects_folder --code FHI-aims
+
+Where ``defects_folder`` is the path to the top level directory containing the defect folder, and is only required if
+different from the current directory.
 
 Instead of a single defect, we can parse the results for **all** defects present
-in a given/current directory using:
+in a given/current directory using the ``-a``/``--all`` flag:
 
 .. code:: bash
 
-    $ snb-parse --all --code vasp --path defects_folder
+    $ snb-parse -a
 
 This generates a ``yaml`` file for each defect, mapping each distortion to the
 final energy of the relaxed structures (in eV). These files are saved to the
@@ -41,12 +49,26 @@ It will generate ``csv`` files for a given/all defects with the final energies
 and structural similarities between the final configurations and a reference one
 (by default the undistorted one). Structural similarity is measured with two metrics:
 a) the sum of atomic displacements between matched sites and b) the maximum distance
-between matched sites. For instance, to analyse the results obtained with ``VASP``
-for the defect ``vac_1_Cd_0``, we can use:
+between matched sites. For instance, to analyse the results obtained with ``VASP`` for a specific defect, we can simply
+run the command ``snb-analyse`` from within the defect directory:
 
 .. code:: bash
 
-    $ snb-analyse --defect vac_1_Cd_0 --code vasp --path defects_folder
+    $ snb-analyse
+
+Again we can alternatively run the command from a different directory and specify which defect to analyse, which code
+was used (if not :code:`VASP`) and which reference structure to use (default = ``Unperturbed``):
+
+.. code:: bash
+
+    $ snb-analyse --defect vac_1_Cd_0 --code FHI-aims --path defects_folder --ref_struct -0.4 --verbose
+
+Again if we want to analyse the results for **all** defects present in a given/current directory, we can use the
+``-a``/``--all`` flag:
+
+.. code:: bash
+
+    $ snb-analyse -a
 
 .. NOTE::
     Further analysis tools are provided through the python API. These are documented in
@@ -58,11 +80,13 @@ for the defect ``vac_1_Cd_0``, we can use:
 Plotting
 -----------
 Energy lowering distortions can be quickly identified by plotting the final energies
-of the relaxed structures versus the distortion factor, using ``snb-plot``:
+of the relaxed structures versus the distortion factor, using ``snb-plot``.
+To plot the results obtained with ``VASP`` for a specific defect, we can simply run the command ``snb-plot`` from within
+the defect directory:
 
 .. code:: bash
 
-    $ snb-plot --defect vac_1_Cd_0 --path defects_folder
+    $ snb-plot
 
 which will generate a figure like the one below:
 
@@ -72,18 +96,32 @@ which will generate a figure like the one below:
 ..
     data from example_results folder
 
-We can make these plots more informative by adding a colorbar measuring the structural
-similarity between the structures:
+We can make these plots more informative by adding a colorbar measuring the structural similarity between the
[…]
