### 2022-10-04 5a32ba5
Update docs to show CLI commands running with default minimal arguments first (to show ease-of-use and user-friendliness etc) then with optional flags for greater control
