### 2022-10-04 b1a741b
Update link in `Analysis.rst` to point to docs API tutorial page rather than GitHub example notebook file
Former-commit-id: 34820ba91f497e42f2dcac89ab62b8f59a0ea81a
