### 2022-10-05 1d5f3b9
Update `snb-generate` tests to use `Int_Cd` rather than `Vac_Cd` so defect site is not (0,0,0) (so fractional and Cartesian coordinates different)
