### 2022-10-05 3aedf5f
Update site-matching routines to be more robust, by now comparing bulk and defect sites within 1 Angstrom radius of specified defect coordinates, and removing any matching ones (site and specie) – warn user if no extra/missing/different species and revert to auto-site-matching. Add tests for this
