+++ tests/test_cli.py
@@ -19,6 +19,7 @@ from pymatgen.core.structure import Structure
 from pymatgen.io.vasp.inputs import Incar, Poscar
 
 from shakenbreak.cli import snb
+from shakenbreak.distortions import rattle
 
 file_path = os.path.dirname(__file__)
 
@@ -271,7 +272,6 @@ class CLITestCase(unittest.TestCase):
 
         # test defect_index option:
         self.tearDown()
-        runner = CliRunner()
         result = runner.invoke(
             snb,
             [
@@ -361,7 +361,6 @@ class CLITestCase(unittest.TestCase):
         # test warning with defect_coords option but wrong site: (matches Cd site in bulk)
         # using Int_Cd because V_Cd is at (0,0,0) so fractional and Cartesian coordinates the same
         self.tearDown()
-        runner = CliRunner()
         with warnings.catch_warnings(record=True) as w:
             result = runner.invoke(
                 snb,
@@ -398,20 +397,101 @@ class CLITestCase(unittest.TestCase):
             self.assertIn(
                 f"\tDefect Site Index / Frac Coords: 65\n"
                 + "            Original Neighbour Distances: [(2.71, 10, 'Cd'), (2.71, 22, 'Cd')]\n"
-                + "            Distorted Neighbour Distances:\n\t[(1.09, 10, 'Cd'), (1.09, 22, 'Cd')]",
+                + "            Distorted Neighbour Distances:\n\t[(1.09, 10, 'Cd'), (1.09, 22, "
+                "'Cd')]",
                 result.output,
             )
             self.assertEqual(
                 Structure.from_file("Int_Cd_mult128_0/Bond_Distortion_-60.0%/POSCAR"),
                 self.Int_Cd_2_minus0pt6_struc_rattled,
             )
-        self.tearDown()
 
-        # test defect_coords working even when slightly off correct site and using slightly rattled
-        # bulk
-        # TODO: ...
+        # test defect_coords working even when slightly off correct site
         self.tearDown()
-        runner = CliRunner()
+        with warnings.catch_warnings(record=True) as w:
+            result = runner.invoke(
+                snb,
+                [
+                    "generate",
+                    "-d",
+                    f"{self.VASP_CDTE_DATA_DIR}/CdTe_Int_Cd_2_POSCAR",
+                    "-b",
+                    f"{self.VASP_CDTE_DATA_DIR}/CdTe_Bulk_Supercell_POSCAR",
+                    "-c",
+                    "0",
+                    "--defect-coords",
+                    0.8,  # 0.8125,  # actual Int_Cd_2 site
+                    0.15,  # 0.1875,
+                    0.85,  # 0.8125,
+                    "-v",
+                ],
+                catch_exceptions=False,
+            )
+            self.assertEqual(result.exit_code, 0)
+            if w:
+                # Check no problems in identifying the defect site
+                self.assertFalse(any(warning.category == UserWarning for warning in w))
+                self.assertFalse(
+                    any(str(warning.message) == warning_message for warning in w)
+                )
+                self.assertFalse(
+                    any("Coordinates" in str(warning.message) for warning in w)
+                )
+
+            self.assertNotIn(f"Auto site-matching", result.output)
+            self.assertIn("--Distortion -60.0%", result.output)
+            self.assertIn(
+                f"\tDefect Site Index / Frac Coords: 65\n"
+                + "            Original Neighbour Distances: [(2.71, 10, 'Cd'), (2.71, 22, 'Cd')]\n"
+                + "            Distorted Neighbour Distances:\n\t[(1.09, 10, 'Cd'), (1.09, 22, "
+                "'Cd')]",
[…]
