### 2022-10-05 29121ae
Formatting and cleanup
