+++ shakenbreak/cli.py
@@ -78,22 +78,27 @@ def identify_defect(
 
             # there should be one site (including specie identity) which does not match between
             # bulk and defect structures
-            for defect_site in list(max_possible_defect_sites_in_defect_struc):
-                for bulk_site in list(max_possible_defect_sites_in_bulk_struc):
-                    if (
-                        defect_site.distance(bulk_site) < 0.5
-                        and defect_site.specie == bulk_site.specie
-                    ):
-                        if bulk_site in max_possible_defect_sites_in_bulk_struc:
-                            max_possible_defect_sites_in_bulk_struc.remove(bulk_site)
-                        if defect_site in max_possible_defect_sites_in_defect_struc:
-                            max_possible_defect_sites_in_defect_struc.remove(
-                                defect_site
-                            )
+            def _remove_matching_sites(bulk_sites, defect_sites):
+                for defect_site in list(defect_sites):
+                    for bulk_site in list(bulk_sites):
+                        if (
+                            defect_site.distance(bulk_site) < 0.5
+                            and defect_site.specie == bulk_site.specie
+                        ):
+                            if bulk_site in bulk_sites:
+                                bulk_sites.remove(bulk_site)
+                            if defect_site in defect_sites:
+                                defect_sites.remove(defect_site)
+                return bulk_sites, defect_sites
+
+            non_matching_bulk_sites, non_matching_defect_sites = _remove_matching_sites(
+                max_possible_defect_sites_in_bulk_struc,
+                max_possible_defect_sites_in_defect_struc,
+            )
 
             if (
-                len(max_possible_defect_sites_in_bulk_struc) == 0
-                and len(max_possible_defect_sites_in_defect_struc) == 0
+                len(non_matching_bulk_sites) == 0
+                and len(non_matching_defect_sites) == 0
             ):
                 warnings.warn(
                     f"Coordinates {defect_coords} were specified for (auto-determined) "
@@ -106,6 +111,8 @@ def identify_defect(
                 )
 
             else:
+                searched = "bulk or defect"
+                possible_defects = []
                 while site_displacement_tol < 0.8:  # loop over distance tolerances
                     possible_defect_sites_in_bulk_struc = sorted(
                         bulk_structure.get_sites_in_sphere(
@@ -125,11 +132,18 @@ def identify_defect(
                     )
                     if (
                         defect_type == "vacancy"
-                    ):  # defect site should be in bulk structure
-                        possible_defects = possible_defect_sites_in_bulk_struc
+                    ):  # defect site should be in bulk structure but not defect structure
+                        possible_defects, _ = _remove_matching_sites(
+                            possible_defect_sites_in_bulk_struc,
+                            max_possible_defect_sites_in_defect_struc,
+                        )
                         searched = "bulk"
                     else:
-                        possible_defects = possible_defect_sites_in_defect_struc
+                        # defect site should be in defect structure but not bulk structure
+                        _, possible_defects = _remove_matching_sites(
+                            max_possible_defect_sites_in_bulk_struc,
+                            possible_defect_sites_in_defect_struc,
+                        )
                         searched = "defect"
 
                     if len(possible_defects) == 1:
