### 2022-10-05 5e08433
Redefine `bulk_supercell_site` in `generate_defect_dict` so it exactly matches generated defect structure (Lavan found a case where pymatgen randomly moved an atom to its periodic image, which then caused problems with site-matching & determining bulk comp), this fixes it
Former-commit-id: 48f9018151bc3a18813d6a6bc8502e9e4c837d33
