### 2022-10-05 ce0719f
Add catch to warn users if manual defect site specification (coords or index) doesn't match auto site-matching defect, but continue with manual specification
Former-commit-id: f97b8234914996c30cc74a7a33ec9eb0064c1940
