### 2022-10-05 #4817 @editorialbot
[REVIEW]: ShakeNBreak: Navigating the defect configurational landscape
**Submitting author:** @ireaml (<a href="http://orcid.org/0000-0001-7651-0814">Irea Mosquera-Lois</a>)
**Repository:** https://github.com/SMTG-UCL/ShakeNBreak
**Branch with paper.md** (empty if default branch):
**Version:** 22.11.6
