+++ shakenbreak/plotting.py
@@ -181,57 +181,6 @@ def _format_defect_name(
         if dummy_h.is_valid_symbol(two_char_string)
     ]
 
-    pre_vacancy_strings = [
-        "V",
-        "v",
-        "V_",
-        "v_",
-        "Vac",
-        "vac",
-        "Vac_",
-        "vac_",
-        "Va",
-        "va",
-        "Va_",
-        "va_",
-    ]
-    post_vacancy_strings = [
-        "_v",  # but not '_V' as could be vanadium
-        "v",  # but not 'V' as could be vanadium
-        "_vac",
-        "_Vac",
-        "vac",
-        "Vac",
-        "va",
-        "Va",
-        "_va",
-        "_Va",
-    ]
-    pre_interstitial_strings = [
-        "i",  # but not 'I' as could be iodine
-        "i_",  # but not 'I_' as could be iodine
-        "Int",
-        "int",
-        "Int_",
-        "int_",
-        "Inter",
-        "inter",
-        "Inter_",
-        "inter_",
-    ]
-    post_interstitial_strings = [
-        "_i",  # but not '_I' as could be iodine
-        "i",  # but not 'I' as could be iodine
-        "_int",
-        "_Int",
-        "int",
-        "Int",
-        "inter",
-        "Inter",
-        "_inter",
-        "_Inter",
-    ]
-
     def _check_matching_defect_format(
         element, name, pre_def_type_list, post_def_type_list
     ):
@@ -278,77 +227,192 @@ def _format_defect_name(
         else:
             return False, None
 
-    if len(possible_two_character_elements) > 0:
-        if len(possible_two_character_elements) == 1:
-            # vacancy?
-            poss_element = possible_two_character_elements[0]
-            if _check_matching_defect_format(
-                poss_element, pre_charge_name, pre_vacancy_strings, post_vacancy_strings
-            ):
-                defect_name = f"V$_{{{poss_element}}}^{{{charge}}}$"
-            elif _check_matching_defect_format(
-                poss_element,
-                pre_charge_name,
-                pre_interstitial_strings,
-                post_interstitial_strings,
-            ):
-                defect_name = f"{poss_element}$_i^{{{charge}}}$"
+    def _try_vacancy_interstitial_match(element, name):
+        defect_name = None
+        pre_vacancy_strings = [
+            "V",
+            "v",
+            "V_",
+            "v_",
+            "Vac",
+            "vac",
+            "Vac_",
+            "vac_",
+            "Va",
+            "va",
+            "Va_",
+            "va_",
+        ]
+        post_vacancy_strings = [
+            "_v",  # but not '_V' as could be vanadium
+            "v",  # but not 'V' as could be vanadium
+            "_vac",
+            "_Vac",
+            "vac",
+            "Vac",
+            "va",
+            "Va",
+            "_va",
+            "_Va",
+        ]
+        pre_interstitial_strings = [
+            "i",  # but not 'I' as could be iodine
+            "i_",  # but not 'I_' as could be iodine
+            "Int",
+            "int",
+            "Int_",
+            "int_",
+            "Inter",
+            "inter",
+            "Inter_",
+            "inter_",
+        ]
+        post_interstitial_strings = [
+            "_i",  # but not '_I' as could be iodine
+            "i",  # but not 'I' as could be iodine
+            "_int",
+            "_Int",
+            "int",
+            "Int",
+            "inter",
+            "Inter",
+            "_inter",
+            "_Inter",
+        ]
+        if _check_matching_defect_format(
+            element, name, pre_vacancy_strings, post_vacancy_strings
+        ):
+            defect_name = f"V$_{{{element}}}^{{{charge}}}$"
+        elif _check_matching_defect_format(
+            element,
[…]
