### 2022-10-06 ea0210f
All-encompassing defect name determination
