+++ shakenbreak/plotting.py
@@ -135,7 +135,8 @@ def _format_defect_name(
     include_site_num_in_name: bool,
 ) -> str:
     """
-    Format defect name for plot titles. (i.e. from vac_1_Cd_0 to $V_{Cd}^{0}$)
+    Format defect name for plot titles. (i.e. from vac_1_Cd_0 to $V_{Cd}^{0}$).
+    Note this assumes "V_" means vacancy not Vanadium.
 
     Args:
         defect_species (:obj:`str`):
@@ -164,22 +165,72 @@ def _format_defect_name(
     if charge > 0:
         charge = "+" + str(charge)  # show positive charges with a + sign
 
-    defect_name = ""
-    dummy_h = Element("H")
-    pre_charge_name = defect_species.rsplit("_", 1)[
-        0
-    ]  # defect name without charge state
-
-    two_character_pairs_in_name = [
-        pre_charge_name[i : i + 2]
-        for i in range(0, len(pre_charge_name), 1)
-        if pre_charge_name[i : i + 3] != "Int" and len(pre_charge_name[i : i + 2]) == 2
-    ]  # need to avoid "Int" giving "In" as valid element
-    possible_two_character_elements = [
-        two_char_string
-        for two_char_string in two_character_pairs_in_name
-        if dummy_h.is_valid_symbol(two_char_string)
-    ]
+    recognised_pre_vacancy_strings = sorted(
+        [
+            "V",
+            "v",
+            "V_",
+            "v_",
+            "Vac",
+            "vac",
+            "Vac_",
+            "vac_",
+            "Va",
+            "va",
+            "Va_",
+            "va_",
+        ],
+        key=len,
+        reverse=True,
+    )
+    recognised_post_vacancy_strings = sorted(
+        [
+            "_v",  # but not '_V' as could be vanadium
+            "v",  # but not 'V' as could be vanadium
+            "_vac",
+            "_Vac",
+            "vac",
+            "Vac",
+            "va",
+            "Va",
+            "_va",
+            "_Va",
+        ],
+        key=len,
+        reverse=True,
+    )
+    recognised_pre_interstitial_strings = sorted(
+        [
+            "i",  # but not 'I' as could be iodine
+            "i_",  # but not 'I_' as could be iodine
+            "Int",
+            "int",
+            "Int_",
+            "int_",
+            "Inter",
+            "inter",
+            "Inter_",
+            "inter_",
+        ],
+        key=len,
+        reverse=True,
+    )
+    recognised_post_interstitial_strings = sorted(
+        [
+            "_i",  # but not '_I' as could be iodine
+            "i",  # but not 'I' as could be iodine
+            "_int",
+            "_Int",
+            "int",
+            "Int",
+            "inter",
+            "Inter",
+            "_inter",
+            "_Inter",
+        ],
+        key=len,
+        reverse=True,
+    )
 
     def _check_matching_defect_format(
         element, name, pre_def_type_list, post_def_type_list
@@ -214,71 +265,38 @@ def _format_defect_name(
                 return True, items[1]
             elif any(
                 (
-                    f"{element}{items[1]}{post_def_type}" in pre_charge_name
+                    f"{element}{items[1]}{post_def_type}" in name
                     for post_def_type in post_def_type_list
                 )
                 or (
-                    f"{element}{items[1]}_{post_def_type}" in pre_charge_name
+                    f"{element}{items[1]}_{post_def_type}" in name
                     for post_def_type in post_def_type_list
                 )
             ):
                 return True, items[1]
+            else:
+                return False, None
 
         else:
             return False, None
 
-    def _try_vacancy_interstitial_match(element, name):
+    def _try_vacancy_interstitial_match(
+        element,
[…]
