### 2022-10-06 c0da9c8
Add tests for encompassing defect naming
