+++ tests/test_plotting.py
@@ -111,7 +111,7 @@ class PlottingDefectsTestCase(unittest.TestCase):
         )
         formatted_ax = plotting._format_axis(
             ax=ax,
-            defect_name="V$_{Cd}^{0}$",
+            defect_name="$V_{Cd}^{0}$",
             y_label="Energy (eV)",
             num_nearest_neighbours=2,
             neighbour_atom="Te",
@@ -119,7 +119,7 @@ class PlottingDefectsTestCase(unittest.TestCase):
         self.assertEqual(formatted_ax.yaxis.get_label().get_text(), "Energy (eV)")
         self.assertEqual(
             formatted_ax.xaxis.get_label().get_text(),
-            "Bond Distortion Factor (for 2 Te near V$_{Cd}^{0}$)",
+            "Bond Distortion Factor (for 2 Te near $V_{Cd}^{0}$)",
         )
         self.assertEqual(
             len(formatted_ax.yaxis.get_ticklabels()), 6 + 2
@@ -134,14 +134,14 @@ class PlottingDefectsTestCase(unittest.TestCase):
         )
         formatted_ax = plotting._format_axis(
             ax=ax,
-            defect_name="V$_{Cd}^{0}$",
+            defect_name="$V_{Cd}^{0}$",
             y_label="Energy (eV)",
             num_nearest_neighbours=2,
             neighbour_atom=None,
         )
         self.assertEqual(
             formatted_ax.xaxis.get_label().get_text(),
-            "Bond Distortion Factor (for 2 NN near V$_{Cd}^{0}$)",
+            "Bond Distortion Factor (for 2 NN near $V_{Cd}^{0}$)",
         )
         # check x label if no defect name
         ax.plot(
@@ -169,7 +169,7 @@ class PlottingDefectsTestCase(unittest.TestCase):
         )
         semi_formatted_ax = plotting._format_axis(
             ax=ax,
-            defect_name="V$_{Cd}^{0}$",
+            defect_name="$V_{Cd}^{0}$",
             y_label="Energy (eV)",
             num_nearest_neighbours=2,
             neighbour_atom="Te",
@@ -261,7 +261,7 @@ class PlottingDefectsTestCase(unittest.TestCase):
         )
         semi_formatted_ax = plotting._format_axis(
             ax=ax,
-            defect_name="V$_{Cd}^{0}$",
+            defect_name="$V_{Cd}^{0}$",
             y_label="Energy (eV)",
             num_nearest_neighbours=2,
             neighbour_atom="Te",
@@ -295,13 +295,13 @@ class PlottingDefectsTestCase(unittest.TestCase):
             defect_species="vac_1_Cd_0",
             include_site_num_in_name=False,
         )
-        self.assertEqual(formatted_name, "V$_{Cd}^{0}$")
+        self.assertEqual(formatted_name, "$V_{Cd}^{0}$")
         # test with site number included
         formatted_name = plotting._format_defect_name(
             defect_species="vac_1_Cd_0",
             include_site_num_in_name=True,
         )
-        self.assertEqual(formatted_name, "V$_{Cd_1}^{0}$")
+        self.assertEqual(formatted_name, "$V_{Cd_1}^{0}$")
         # test interstitial case
         formatted_name = plotting._format_defect_name(
             defect_species="Int_Cd_1_0",
@@ -321,15 +321,19 @@ class PlottingDefectsTestCase(unittest.TestCase):
[…]
