### 2022-10-06 70eb05e
Update site number determination for substitutions and add test for this. Make code more efficient
