### 2022-10-06 c6971dd
Put _format_defect_name() calls into try except statements, so if we try plot with an unformattable name (e.g. if parsing and plotting from a renamed folder), we don't break but just don't have formatted defect name on axis
