### 2022-10-07 45dfe40
Add tests for plotting defect with unrecognised file name, giving warning but still plotting fine
