### 2022-10-07 fe08605
Add print statement to debug failing GH Actions test but fine on local?
