+++ tests/test_cli.py
@@ -384,31 +384,28 @@ class CLITestCase(unittest.TestCase):
                 ],
                 catch_exceptions=False,
             )
-            self.assertEqual(result.exit_code, 0)
-            warning_message = (
-                "Coordinates (0.0, 0.0, 0.0) were specified for (auto-determined) interstitial "
-                "defect, but there are no extra/missing/different species within a 2.5 Å radius of "
-                "this site when comparing bulk and defect structures. If you are trying to "
-                "generate non-defect polaronic distortions, please use the distort() and rattle() "
-                "functions in shakenbreak.distortions via the Python API. Reverting to auto-site "
-                "matching instead."
-            )
-            self.assertTrue(any(warning.category == UserWarning for warning in w))
-            self.assertTrue(
-                any(str(warning.message) == warning_message for warning in w)
-            )
-            self.assertIn("--Distortion -60.0%", result.output)
-            self.assertIn(
-                f"\tDefect Site Index / Frac Coords: 65\n"
-                + "            Original Neighbour Distances: [(2.71, 10, 'Cd'), (2.71, 22, 'Cd')]\n"
-                + "            Distorted Neighbour Distances:\n\t[(1.09, 10, 'Cd'), (1.09, 22, "
-                "'Cd')]",
-                result.output,
-            )
-            self.assertEqual(
-                Structure.from_file("Int_Cd_mult128_0/Bond_Distortion_-60.0%/POSCAR"),
-                self.Int_Cd_2_minus0pt6_struc_rattled,
-            )
+        self.assertEqual(result.exit_code, 0)
+        warning_message = (
+            "Coordinates (0.0, 0.0, 0.0) were specified for (auto-determined) interstitial "
+            "defect, but there are no extra/missing/different species within a 2.5 Å radius of "
+            "this site when comparing bulk and defect structures. If you are trying to "
+            "generate non-defect polaronic distortions, please use the distort() and rattle() "
+            "functions in shakenbreak.distortions via the Python API. Reverting to auto-site "
+            "matching instead."
+        )
+        self.assertTrue(any(warning.category == UserWarning for warning in w))
+        self.assertTrue(any(str(warning.message) == warning_message for warning in w))
+        self.assertIn("--Distortion -60.0%", result.output)
+        self.assertIn(
+            f"\tDefect Site Index / Frac Coords: 65\n"
+            + "            Original Neighbour Distances: [(2.71, 10, 'Cd'), (2.71, 22, 'Cd')]\n"
+            + "            Distorted Neighbour Distances:\n\t[(1.09, 10, 'Cd'), (1.09, 22, 'Cd')]",
+            result.output,
+        )
+        self.assertEqual(
+            Structure.from_file("Int_Cd_mult128_0/Bond_Distortion_-60.0%/POSCAR"),
+            self.Int_Cd_2_minus0pt6_struc_rattled,
+        )
 
         # test defect_coords working even when slightly off correct site
         self.tearDown()
@@ -431,77 +428,77 @@ class CLITestCase(unittest.TestCase):
                 ],
                 catch_exceptions=False,
             )
-            self.assertEqual(result.exit_code, 0)
-            if w:
-                # Check no problems in identifying the defect site
-                self.assertFalse(any(warning.category == UserWarning for warning in w))
-                self.assertFalse(
-                    any(str(warning.message) == warning_message for warning in w)
-                )
-                self.assertFalse(
-                    any("Coordinates" in str(warning.message) for warning in w)
-                )
-
-            self.assertNotIn(f"Auto site-matching", result.output)
[…]
