### 2022-10-07 7d0d2c2
Update test plot fake defect as fake defect_species no longer raises error
Former-commit-id: 7f8f81b6fa43a16898af6ba546b8fd0cee4e4250
