### 2022-10-07 7f8f81b
Update test plot fake defect as fake defect_species no longer raises error
