### 2022-10-12 363ded8
Some notebook formatting updates (`input_files` -> `SnB_input_files`, run `plot_all_defects()` twice so first plot isn't small, quieten `write_vasp_files()` output by assigning to `defects_dict`, `distortion_metadata` variables...)
Former-commit-id: 7e4705843ea1f09dfc94f85892240fe3c98748fd
