### 2022-10-12 3556efa
Add bug fix for rare event where the first matching structure when doing pruning structure comparisons for `low_energy_defects` is an imported distortion, which previously threw a KeyError because the `struc_key` had been reformatted to a float so no longer compatible with `defect_structures_dict` and `defect_energies_dict`. Tested and fixed now
