### 2022-10-12 fd20bb4
Fix so that if `energies_file` not correctly parsed for some defects when running `write_groundstate_structure()` (e.g. if only results for vacancies but not substitutions have come in), we don't break with an error but instead give a warning and skip. Tested and fixes this
