### 2022-10-12 d2cd721
Add catch to `SnB_run.sh` not to cp CONTCAR to POSCAR if CONTCAR is empty (i.e. if VASP didn't make a full ionic step), rare case
Former-commit-id: 6c1c7ce22fb557282c016a341050dc901dcc5107
