### 2022-10-13 f075e79
Update `SnB_run.sh`
