### 2022-10-13 d02ac1f
Update `SnB_run.sh`
Former-commit-id: f075e79e6c77c9d9b4ba554850d2db6629fd51a4
