+++ shakenbreak/bash_scripts/SnB_run.sh
@@ -56,7 +56,7 @@ SnB_run_loop () {
           then
           # count number of ionic steps with positive energies, after the first 5 ionic steps
           pos_energies=$( grep entropy= OUTCAR | awk 'FNR>5 && $NF !~ /^-/{print $0}' | wc -l )
-          errors=$(grep -Ec "(EDDDAV|ZHEGV|sick job|SICK JOB|CNORMN|ZPOTRF|ZTRTRI|SICK JOB)" OUTCAR)
+          errors=$(grep -Ec "(EDDDAV|ZHEGV|CNORMN|ZPOTRF|ZTRTRI)" OUTCAR)
           if (( pos_energies > 0 )) ||  (( errors > 0 ))  # if there are positive energies or errors in OUTCAR
             then
             echo "Positive energies or forces error encountered for ${i%/}, ignoring and renaming to ${i%/}_High_Energy"
