### 2022-10-13 f86d496
Add SnB CLI GIF to readme and docs
Former-commit-id: b831e75e35355ae1c3487cdf4dc95a506698fbea
