### 2022-10-13 b831e75
Add SnB CLI GIF to readme and docs
