### 2022-10-13 93f5675
Point to CLI summary gif at top of CLI tutorial page
