+++ docs/conf.py
@@ -25,7 +25,7 @@ copyright = '2022, Irea Mosquera-Lois, Seán R. Kavanagh'
 author = 'Irea Mosquera-Lois, Seán R. Kavanagh'
 
 # The full version, including alpha/beta/rc tags
-release = '22.9.21'
+release = '22.10.13'
 
 
 # -- General configuration ---------------------------------------------------
