### 2022-10-13 54d90e7
check if symbols in `distorted_element` dict are valid element symbols
