### 2022-10-13 f333e56
refactor `_setup_distorted_defect_dict` to use Defect object
Former-commit-id: 71f6a550390a739372783eff6748bfdf6a8136cf
