### 2022-10-13 71f6a55
refactor `_setup_distorted_defect_dict` to use Defect object
