### 2022-10-14 5a7bc5a
Refactor testing in workflows to do loop over files (as done with `build_and_test.yml`), seems to avoid OOM error with exit code 137 (currently failing)
Former-commit-id: 8585dc66d18cb5a3cc0a6fe51cd9a922d9c12971
