### 2022-10-14 8585dc6
Refactor testing in workflows to do loop over files (as done with `build_and_test.yml`), seems to avoid OOM error with exit code 137 (currently failing)
