+++ .github/workflows/release.yml
@@ -28,7 +28,11 @@ jobs:
 
       - name: Test
         run:  |
-          pytest --ignore=tests/test_local.py  # Ignore local tests file, which tests INCAR and POTCAR file writing but not possible on GitHub Actions
+          for test in tests/test_*py
+            do if [[ "$test" != *"test_local"* ]]
+              then pytest $test
+            fi
+          done  # Ignore local tests file, which tests INCAR and POTCAR file writing but not possible on GitHub Actions
           pytest --mpl tests/test_shakenbreak.py  # test output plots
           pytest --mpl tests/test_plotting.py  # test output plots
 
