### 2022-10-14 575ad7c
Add 10 minute sleep command to `pip_install_test.yml` as PyPI often hasn't updated to the most recent released version quick enough, so GH Actions ends up downloading the older version and can then fail some tests. Checked and this works now
Former-commit-id: 53d08dd9b1ef758b8a74fac7dbed2ba78cff869d
