### 2022-10-14 5cc3f39
fix broken links to example config files in cli tutorial
Former-commit-id: 5c7749815da9534dbbb60a8531049df38a0a1e0e
