### 2022-10-14 e02bd70
add `from_dicts` method to generate Distortions from doped/PyCDT defect_dicts (for compatibility with older snb and pymatgen versions)
