+++ shakenbreak/input.py
@@ -759,13 +759,6 @@ def apply_snb_distortions(
     return distorted_defect_dict
 
 
-# TODO: As discussed, need to refactor Distortions() to take a dictionary of Defect() objects
-# rather than a dict of (doped/pmg/pycdt) defect_dict
-# We need a way to store the defect charge states defined by the user though, as with
-# the get_charge_states() method we need to give the padding
-# a)
-
-
 class Distortions:
     """
     Class to apply rattle and bond distortion to all defects in `defects_dict`
@@ -1862,6 +1855,129 @@ class Distortions:
                         )  # write parameters file
 
         return distorted_defects_dict, self.distortion_metadata
+
+    @classmethod
+    def from_dict(
+        cls,
+        doped_defects_dict: dict,
+        oxidation_states: Optional[dict] = None,
+        dict_number_electrons_user: Optional[dict] = None,
+        distortion_increment: float = 0.1,
+        bond_distortions: Optional[list] = None,
+        local_rattle: bool = False,
+        stdev: float = 0.25,
+        distorted_elements: Optional[dict] = None,
+        **kwargs,  # for mc rattle
+    ) -> None:
+        """
+        Initialise Distortion() class using a dictionary of DOPED/PyCDT defect dicts
+        (instead of pymatgen-analysis-defects Defect() objects).
+
+        Args:
+            defects_dict (:obj:`dict`):
+                Dictionary of DOPED/pyCDT defect dictionaries.
+            oxidation_states (:obj:`dict`):
+                Dictionary of oxidation states for species in your material,
+                used to determine the number of defect neighbours to distort
+                (e.g {"Cd": +2, "Te": -2}). If none is provided, the oxidation
+                states will be guessed based on the bulk composition and most
+                common oxidation states of any extrinsic species.
+            dict_number_electrons_user (:obj:`dict`):
+                Optional argument to set the number of extra/missing charge
+                (negative of electron count change) for the input defects
+                in their neutral state, as a dictionary with format
+                {'defect_name': charge_change} where charge_change is the
+                negative of the number of extra/missing electrons.
+                (Default: None)
+            distortion_increment (:obj:`float`):
+                Bond distortion increment. Distortion factors will range from
+                0 to +/-0.6, in increments of `distortion_increment`.
+                Recommended values: 0.1-0.3
+                (Default: 0.1)
+            bond_distortions (:obj:`list`):
+                List of bond distortions to apply to nearest neighbours,
+                instead of the default set (e.g. [-0.5, 0.5]).
+                (Default: None)
+            local_rattle (:obj:`bool`):
+                Whether to apply random displacements that tail off as we move
+                away from the defect site. Not recommended as typically worsens
+                performance. If False (default), all supercell sites are rattled
+                with the same amplitude (full rattle).
+                (Default: False)
+            stdev (:obj:`float`):
+                Standard deviation (in Angstroms) of the Gaussian distribution
+                from which random atomic displacement distances are drawn during
+                rattling. Recommended values: 0.25, or 0.15 for strongly-bound
+                /ionic materials.
+                (Default: 0.25)
+            distorted_elements (:obj:`dict`):
+                Optional argument to specify the neighbouring elements to
+                distort for each defect, in the form of a dictionary with
+                format {'defect_name': ['element1', 'element2', ...]}
+                (e.g {'vac_1_Cd': ['Te']}). If None, the closest neighbours to
+                the defect are chosen.
+                (Default: None)
+            **kwargs:
+                Additional keyword arguments to pass to `hiphive`'s
+                `mc_rattle` function. These include:
+                - d_min (:obj:`float`):
+                    Minimum interatomic distance (in Angstroms). Monte Carlo rattle
[…]
