### 2022-10-14 4472545
add `from_dicts` method to generate Distortions from doped/PyCDT defect_dicts (for compatibility with older snb and pymatgen versions)
Former-commit-id: e02bd705285d4f46f28295318f363d9fb80379f6
