### 2022-10-14 e5c691c
update apply_distortions
Former-commit-id: 38a4d865ee6d08f06bf0457b3c900899db235490
