### 2022-10-14 8611021
update cli module to use Defect() class
