### 2022-10-14 8485bf4
update cli module to use Defect() class
Former-commit-id: 8611021474794e5dfeee02d7e3d05e2be57f7b79
