### 2022-10-15 1e97c51
use different names for symmetry inequivalent defects (pymatgen-analysis defects names them equally (e.g. v_Sb))
Former-commit-id: 7c57d9d43dd45077e8c70f8b96e7e8d83d5a3003
