### 2022-10-15 6f54586
add test for `generate_defect_object`
Former-commit-id: 0d5e6da72cc473774ff0216bbed53b0209f4b6ae
