### 2022-10-15 35f36cd
make test for `generate_defect_object` more robust
