### 2022-10-15 e0159b9
make test for `generate_defect_object` more robust
Former-commit-id: 35f36cdd1367f2450d39bf11f687211b25440eea
