### 2022-10-16 af6a39a
update `generate` test to work with Defect class
Former-commit-id: 2f66144a7f5a4fd6268fcdf8cd088bd677062fd2
