+++ tests/test_cli.py
@@ -190,7 +190,7 @@ class CLITestCase(unittest.TestCase):
         self.assertEqual(result.exit_code, 0)
         self.assertIn(
             f"Auto site-matching identified {self.VASP_CDTE_DATA_DIR}/CdTe_V_Cd_POSCAR "
-            f"to be type Vacancy with site Cd at [0.000, 0.000, 0.000]",
+            f"to be type Vacancy with site Cd2+ at [0.000, 0.000, 0.000]",
             result.output,
         )
         self.assertIn(
@@ -206,10 +206,11 @@ class CLITestCase(unittest.TestCase):
             "'0.3', '0.4', '0.5', '0.6']. Then, will rattle with a std dev of 0.25 Å",
             result.output,
         )
-        self.assertIn("Defect: Vac_Cd_mult32", result.output)
+        defect_name = "v_Cd"
+        self.assertIn(f"Defect: {defect_name}", result.output)
         self.assertIn("Number of missing electrons in neutral state: 2", result.output)
         self.assertIn(
-            "Defect Vac_Cd_mult32 in charge state: 0. Number of distorted neighbours: 2",
+            f"Defect {defect_name} in charge state: 0. Number of distorted neighbours: 2",
             result.output,
         )
         self.assertIn("--Distortion -60.0%", result.output)
@@ -228,14 +229,14 @@ class CLITestCase(unittest.TestCase):
         )
 
         # check if correct files were created:
-        V_Cd_Bond_Distortion_folder = "Vac_Cd_mult32_0/Bond_Distortion_-50.0%"
+        V_Cd_Bond_Distortion_folder = f"{defect_name}_0/Bond_Distortion_-50.0%"
         self.assertTrue(os.path.exists(V_Cd_Bond_Distortion_folder))
         V_Cd_minus0pt5_rattled_POSCAR = Poscar.from_file(
             V_Cd_Bond_Distortion_folder + "/POSCAR"
         )
         self.assertEqual(
             V_Cd_minus0pt5_rattled_POSCAR.comment,
-            "-50.0%__num_neighbours=2__Vac_Cd_mult32",
+            f"-50.0%__num_neighbours=2__{defect_name}",
         )  # default
         self.assertEqual(
             V_Cd_minus0pt5_rattled_POSCAR.structure,
@@ -243,7 +244,7 @@ class CLITestCase(unittest.TestCase):
         )
 
         # Test recognises distortion_metadata.json:
-        if_present_rm("Vac_Cd_mult32_0")  # but distortion_metadata.json still present
+        if_present_rm(f"{defect_name}_0")  # but distortion_metadata.json still present
         runner = CliRunner()
         result = runner.invoke(
             snb,
@@ -262,7 +263,7 @@ class CLITestCase(unittest.TestCase):
         self.assertNotIn(
             f"Auto site-matching identified"
             f" {self.VASP_CDTE_DATA_DIR}/CdTe_V_Cd_POSCAR "
-            f"to be type Vacancy with site Cd at [0.000, 0.000, 0.000]",
+            f"to be type Vacancy with site Cd2+ at [0.000, 0.000, 0.000]",
             result.output,
         )
         self.assertIn(
@@ -277,10 +278,10 @@ class CLITestCase(unittest.TestCase):
             "'0.3', '0.4', '0.5', '0.6']. Then, will rattle with a std dev of 0.25 Å",
             result.output,
         )
-        self.assertIn("Defect: Vac_Cd_mult32", result.output)
+        self.assertIn(f"Defect: {defect_name}", result.output)
         self.assertIn("Number of missing electrons in neutral state: 2", result.output)
         self.assertIn(
-            "Defect Vac_Cd_mult32 in charge state: 0. Number of distorted neighbours: 2",
[…]
