### 2022-10-16 2f66144
update `generate` test to work with Defect class
