### 2022-10-16 2cb385c
refactor `test_generate_config` to use Defect() class
Former-commit-id: 84b47b8e9421dea9cfb5f067fb2c00c104e4a93a
