### 2022-10-16 84b47b8
refactor `test_generate_config` to use Defect() class
