### 2022-10-16 88c719d
update tear down with pymatgen-analysis-defects naming
Former-commit-id: 70ab88213bac7b378f854d834b58ac6602e66a61
