### 2022-10-16 e53e1ec
refactor test_get_bulk_comp
Former-commit-id: f5dfad457bc94cecb7f98b085aac69509896e343
