### 2022-10-16 7c25134
refactor test_calc_number_electrons
Former-commit-id: de2adcf4a4a0a417193d7daa0db5062b16cb56e9
