### 2022-10-16 8478550
refactor test_apply_rattle_bond_distortions
Former-commit-id: 84d1856a54b94682bce7c6f645d75b55061faff3
