+++ tests/test_input.py
@@ -10,6 +10,7 @@ from unittest.mock import patch
 
 import numpy as np
 from ase.calculators.aims import Aims
+from pymatgen.analysis.defects.core import StructureMatcher
 from pymatgen.core.structure import Composition, PeriodicSite, Structure
 from pymatgen.io.vasp.inputs import Poscar
 
@@ -68,7 +69,7 @@ class InputTestCase(unittest.TestCase):
         # Refactor to Defect() objects
         self.V_Cd = cli.generate_defect_object(self.V_Cd_dict, self.cdte_defect_dict["bulk"])
         self.Int_Cd_2 = cli.generate_defect_object(self.Int_Cd_2_dict, self.cdte_defect_dict["bulk"])
-        # Setup structures
+        # Setup structures and add oxidation states (as pymatgen-analysis-defects does it)
         self.V_Cd_struc = Structure.from_file(
             os.path.join(self.VASP_CDTE_DATA_DIR, "CdTe_V_Cd_POSCAR")
         )
@@ -86,11 +87,6 @@ class InputTestCase(unittest.TestCase):
         self.V_Cd_minus0pt5_struc_kwarged = Structure.from_file(
             os.path.join(self.VASP_CDTE_DATA_DIR, "CdTe_V_Cd_-50%_Kwarged_POSCAR")
         )
-        self.V_Cd_distortion_parameters = {
-            "unique_site": np.array([0.0, 0.0, 0.0]),
-            "num_distorted_neighbours": 2,
-            "distorted_atoms": [(33, "Te"), (42, "Te")],
-        }
         self.Int_Cd_2_struc = Structure.from_file(
             os.path.join(self.VASP_CDTE_DATA_DIR, "CdTe_Int_Cd_2_POSCAR")
         )
@@ -104,6 +100,22 @@ class InputTestCase(unittest.TestCase):
                 self.VASP_CDTE_DATA_DIR, "CdTe_Int_Cd_2_-60%_Distortion_NN_10_POSCAR"
             )
         )
+        # for struct in [
+        #     self.V_Cd_struc,
+        #     self.V_Cd_minus0pt5_struc_rattled,
+        #     self.V_Cd_minus0pt5_struc_0pt1_rattled,
+        #     self.Int_Cd_2_struc,
+        #     self.Int_Cd_2_minus0pt6_struc_rattled,
+        #     self.Int_Cd_2_minus0pt6_NN_10_struc_unrattled,
+        # ]:
+        #     struct.add_oxidation_state_by_guess()
+
+        # Setup distortion parameters
+        self.V_Cd_distortion_parameters = {
+            "unique_site": np.array([0.0, 0.0, 0.0]),
+            "num_distorted_neighbours": 2,
+            "distorted_atoms": [(33, "Te"), (42, "Te")],
+        }
         self.Int_Cd_2_normal_distortion_parameters = {
             "unique_site": self.Int_Cd_2_dict["unique_site"].frac_coords,
             "num_distorted_neighbours": 2,
@@ -290,7 +302,7 @@ class InputTestCase(unittest.TestCase):
     def test_apply_rattle_bond_distortions_V_Cd(self):
         """Test _apply_rattle_bond_distortions function for V_Cd"""
         V_Cd_distorted_dict = input._apply_rattle_bond_distortions(
-            self.V_Cd_dict,
+            self.V_Cd,
             num_nearest_neighbours=2,
             distortion_factor=0.5,
         )
@@ -314,6 +326,9 @@ class InputTestCase(unittest.TestCase):
             d_min=d_min,
             active_atoms=rattling_atom_indices,
         )
+        # pymatgen-analysis-defects decorates the structure with oxidation states.
[…]
