### 2022-10-16 1435d10
refactor `test_apply_snb_distortions`
