+++ tests/test_input.py
@@ -119,25 +119,27 @@ class InputTestCase(unittest.TestCase):
         self.Int_Cd_2_normal_distortion_parameters = {
             "unique_site": self.Int_Cd_2_dict["unique_site"].frac_coords,
             "num_distorted_neighbours": 2,
-            "distorted_atoms": [(10, "Cd"), (22, "Cd")],
-            "defect_site_index": 65,
+            "distorted_atoms": [(10+1, "Cd"), (22+1, "Cd")],  # +1 because
+            # interstitial is added at the beggining of the structure
+            "defect_site_index": 1,
         }
         self.Int_Cd_2_NN_10_distortion_parameters = {
             "unique_site": self.Int_Cd_2_dict["unique_site"].frac_coords,
             "num_distorted_neighbours": 10,
             "distorted_atoms": [
-                (10, "Cd"),
-                (22, "Cd"),
-                (29, "Cd"),
-                (1, "Cd"),
-                (14, "Cd"),
-                (24, "Cd"),
-                (30, "Cd"),
-                (38, "Te"),
-                (54, "Te"),
-                (62, "Te"),
+                (10+1, "Cd"),
+                (22+1, "Cd"),
+                (29+1, "Cd"),
+                (1+1, "Cd"),
+                (14+1, "Cd"),
+                (24+1, "Cd"),
+                (30+1, "Cd"),
+                (38+1, "Te"),
+                (54+1, "Te"),
+                (62+1, "Te"),
+                # +1 because interstitial is added at the beggining of the structure
             ],
-            "defect_site_index": 65,
+            "defect_site_index": 1,
         }
 
         # Note that Int_Cd_2 has been chosen as a test case, because the first few nonzero bond
@@ -476,16 +478,17 @@ class InputTestCase(unittest.TestCase):
             stdev=0.25,
             verbose=True,
         )
-        self.assertDictEqual(self.V_Cd_dict, V_Cd_distorted_dict["Unperturbed"])
+        self.assertEqual(self.V_Cd, V_Cd_distorted_dict["Unperturbed"])
 
         distorted_V_Cd_struc = V_Cd_distorted_dict["distortions"][
             "Bond_Distortion_-50.0%"
         ]
+        distorted_V_Cd_struc.remove_oxidation_states()  # pymatgen-analysis-defects add ox. states
         self.assertNotEqual(self.V_Cd_struc, distorted_V_Cd_struc)
         self.assertEqual(self.V_Cd_minus0pt5_struc_rattled, distorted_V_Cd_struc)
 
         V_Cd_0pt1_distorted_dict = input.apply_snb_distortions(
-            self.V_Cd_dict,
+            self.V_Cd,
             num_nearest_neighbours=2,
             bond_distortions=[-0.5],
             stdev=0.1,
@@ -494,6 +497,7 @@ class InputTestCase(unittest.TestCase):
         distorted_V_Cd_struc = V_Cd_0pt1_distorted_dict["distortions"][
             "Bond_Distortion_-50.0%"
         ]
+        distorted_V_Cd_struc.remove_oxidation_states()
         self.assertNotEqual(self.V_Cd_struc, distorted_V_Cd_struc)
         self.assertEqual(self.V_Cd_minus0pt5_struc_0pt1_rattled, distorted_V_Cd_struc)
 
@@ -503,7 +507,7 @@ class InputTestCase(unittest.TestCase):
         )
 
[…]
