### 2022-10-16 c3ec16e
Add note about strongly-correlated / magnetic materials being v sensitive to rattling, need small `stdev`
