### 2022-10-16 9d84210
refactor test_Distortions_initialisation
