### 2022-10-16 068dc36
refactor test_Distortions_initialisation
Former-commit-id: 9d84210ff5f497b3292d1ace758363869cdd4902
