### 2022-10-16 661ea15
refactor test_write_vasp_files
