+++ shakenbreak/input.py
@@ -1277,31 +1277,10 @@ class Distortions:
 
         distorted_defects_dict = {}  # Store distorted & undistorted structures
 
-        comb_defs = functools.reduce(
-            lambda x, y: x + y,
-            [self.defects_dict[key] for key in self.defects_dict if key != "bulk"],
-        )
-        # To allow user to specify defect names (with CLI), defect_dict can be either
-        # a dict of lists or a dict of dicts (e.g. {"vacancies": {"name_1": Vacancy()}})
-        # To account for this, here we refactor the list into a dict
-        if isinstance(comb_defs, list):
-            # Check if defect names are repeated
-            names = [defect.name for defect in comb_defs]
-            counter = Counter(names)
-            if any([value > 1 for value in counter.values()]):
-                # Give different names to symmetry inequivalent defects.
-                # Naming convention is f"{defectname}_s{siteindex}" (e.g. v_Cd_s0)
-                comb_defs = {
-                    f"{defect.name}_s{defect.defect_site_index}": defect
-                    for defect in comb_defs
-                }
-                print(
-                    "There are symmetry inequivalent defects."
-                    " To avoid using the same name for them, the names will be refactored"
-                    " as {defect_name}_s{defect_site_index} (e.g. v_Cd_s0)"
-                )
-            else:
-                comb_defs = {defect.name: defect for defect in comb_defs}
+        # Remove vacancies/substitutions/interstitials classification
+        comb_defs = {}
+        for defect_dict in self.defects_dict.values():
+            comb_defs.update(defect_dict)
 
         for defect_name, defect_object in comb_defs.items():  # loop for each defect
             bulk_supercell_site = defect_object.site
