### 2022-10-16 1834978
refactor test_write_vasp_files
Former-commit-id: 661ea157ec1ae7a2335aac2b5c24c4f7b7bd416b
