### 2022-10-16 8e3bbe3
refactor tests for  `write_cp2k_files`, `write_espresso_files`, `write_fhi_aims_files`, `write_castep_files`
Former-commit-id: 4140c08a216f4bea9053454b943960b13b36c0c3
