### 2022-10-16 0a9c490
refactor local rattle test
