+++ tests/test_input.py
@@ -1598,7 +1598,7 @@ class InputTestCase(unittest.TestCase):
         # check files are not written if `apply_distortions()` method is used
         for i in self.cdte_defect_folders:
             if_present_rm(i)  # remove test-generated defect folders
-        reduced_V_Cd = self.V_Cd.copy()
+        reduced_V_Cd = copy.copy(self.V_Cd)
         reduced_V_Cd.user_charges = [0]
         oxidation_states = {"Cd": +2, "Te": -2}
         bond_distortions = list(np.arange(-0.6, 0.601, 0.05))
@@ -1701,8 +1701,10 @@ class InputTestCase(unittest.TestCase):
         self.assertTrue("Bond_Distortion_-75.0%" in V_Cd_distortions_dict)
 
         # test short interatomic distance distortions not omitted when Hydrogen knocking about
-        fake_hydrogen_V_Cd_dict = reduced_V_Cd_dict.copy()
-        fake_hydrogen_V_Cd_dict["supercell"]["structure"][0].species = "H"
+        # TODO: Check why the H atom gets removed when the Defect() is created!?
+        fake_hydrogen_V_Cd_dict = copy.copy(self.V_Cd_dict)
+        fake_hydrogen_V_Cd_dict["charges"] = [0]
+        fake_hydrogen_V_Cd_dict["supercell"]["structure"][4].species = "H"
         fake_hydrogen_V_Cd = cli.generate_defect_object(
             fake_hydrogen_V_Cd_dict,
             self.cdte_defect_dict["bulk"]
@@ -1732,7 +1734,7 @@ class InputTestCase(unittest.TestCase):
         )
         V_Cd_distortions_dict = distortion_defect_dict["vac_1_Cd"]["charges"][0]["structures"][
             "distortions"]
-        self.assertEqual(len(V_Cd_distortions_dict), 21)
+        self.assertEqual(len(V_Cd_distortions_dict), 21)  # 21 total distortions
         self.assertTrue("Bond_Distortion_-80.0%" in V_Cd_distortions_dict)
         self.assertTrue("Bond_Distortion_-75.0%" in V_Cd_distortions_dict)
 
@@ -1740,8 +1742,8 @@ class InputTestCase(unittest.TestCase):
         self,
     ):
         """ "Test option local_rattle of Distortions class"""
-        reduced_V_Cd_dict = self.V_Cd.copy()
-        reduced_V_Cd_dict.user_charges = [0]
+        reduced_V_Cd = copy.copy(self.V_Cd)
+        reduced_V_Cd.user_charges = [0]
         oxidation_states = {"Cd": +2, "Te": -2}
         dist = input.Distortions(
             {"vacancies": {"vac_1_Cd": reduced_V_Cd}},
@@ -1752,19 +1754,21 @@ class InputTestCase(unittest.TestCase):
         self.assertTrue(dist.local_rattle)
         defects_dict, metadata_dict = dist.apply_distortions()
         # Check structure
+        generated_struct = defects_dict["vac_1_Cd"]["charges"][0]["structures"]["distortions"][
+                "Bond_Distortion_-30.0%"
+            ]
+        generated_struct.remove_oxidation_states()
         self.assertEqual(
             Structure.from_file(
                 f"{self.VASP_CDTE_DATA_DIR}/vac_1_Cd_0_-30.0%_Distortion_tailed_off_rattle_POSCAR"
             ),
-            defects_dict["vac_1_Cd"]["charges"][0]["structures"]["distortions"][
-                "Bond_Distortion_-30.0%"
-            ],
+            generated_struct,
         )
         # Check if option written to metadata file
         self.assertTrue(metadata_dict["distortion_parameters"]["local_rattle"])
 
[…]
