### 2022-10-17 bea7722
refactor integration tets
Former-commit-id: dd622eb679ea825b6be8f5a0c71a8c01f6c55795
