### 2022-10-18 ce7c373
Parse calculation energies even if not converged (but still warn!), and use directory name rather than "Label_not_Recognized" if name can't be parsed (more informative). Add test for this
