+++ shakenbreak/io.py
@@ -106,86 +106,101 @@ def parse_energies(
             dumpfn(energies, filename)
 
     def parse_vasp_energy(defect_dir, dist, energy, outcar):
-        # regrep faster than using Outcar/vasprun class
+        """Parse VASP energy from OUTCAR file."""
+        converged = False
         if os.path.exists(os.path.join(defect_dir, dist, "OUTCAR")):
             outcar = os.path.join(defect_dir, dist, "OUTCAR")
-        if outcar:
-            if _match(
+        if outcar:  # regrep faster than using Outcar/vasprun class
+            energy = _match(outcar, r"energy\(sigma->0\)\s+=\s+([\d\-\.]+)")[0][0][
+                0
+            ]  # Energy of first match
+            converged = _match(
                 outcar, "required accuracy"
-            ):  # check if ionic relaxation is converged
-                energy = _match(outcar, r"energy\(sigma->0\)\s+=\s+([\d\-\.]+)")[0][0][
-                    0
-                ]  # Energy of first match
-        return energy, outcar
+            )  # check if ionic relaxation converged
+        return converged, energy, outcar
 
-    def parse_espresso_energy(defect_dir, dist, energy, outcar):
+    def parse_espresso_energy(defect_dir, dist, energy, espresso_out):
+        """Parse Quantum Espresso energy from espresso.out file."""
         if os.path.join(
             defect_dir, dist, "espresso.out"
         ):  # Default SnB output filename
-            outcar = os.path.join(defect_dir, dist, "espresso.out")
+            espresso_out = os.path.join(defect_dir, dist, "espresso.out")
         elif os.path.exists(os.path.join(defect_dir, dist, filename)):
-            outcar = os.path.join(defect_dir, dist, filename)
-        if outcar:
-            energy_in_Ry = _match(outcar, r"!    total energy\s+=\s+([\d\-\.]+)")
+            espresso_out = os.path.join(defect_dir, dist, filename)
+        if espresso_out:
+            energy_in_Ry = _match(espresso_out, r"!    total energy\s+=\s+([\d\-\.]+)")
             if energy_in_Ry:
                 # Energy of first match, in Rydberg
                 energy = float(Energy(float(energy_in_Ry[0][0][0]), "Ry").to("eV"))
-        return energy, outcar
+        return True, energy, espresso_out  # no check for ionic convergence in QE
 
-    def parse_cp2k_energy(defect_dir, dist, energy, outcar):
+    def parse_cp2k_energy(defect_dir, dist, energy, cp2k_out):
+        """Parse CP2K energy from relax.out file."""
+        converged = False
         if os.path.join(defect_dir, dist, "relax.out"):
-            outcar = os.path.join(defect_dir, dist, "relax.out")
+            cp2k_out = os.path.join(defect_dir, dist, "relax.out")
         elif os.path.exists(os.path.join(defect_dir, dist, filename)):
-            outcar = os.path.join(defect_dir, dist, filename)
-        if outcar and _match(
-            outcar, "GEOMETRY OPTIMIZATION COMPLETED"
-        ):  # check if ionic relaxation is converged
-            energy_in_Ha = _match(outcar, r"Total energy:\s+([\d\-\.]+)")
-            if energy_in_Ha:
-                # Energy of first match in Hartree
-                energy = float(Energy(energy_in_Ha[0][0][0], "Ha").to("eV"))
-        return energy, outcar
+            cp2k_out = os.path.join(defect_dir, dist, filename)
+        if cp2k_out:
+            converged = _match(
+                cp2k_out, "GEOMETRY OPTIMIZATION COMPLETED"
+            )  # check if ionic
+            # relaxation is converged
[…]
