### 2022-10-18 876970f
Parse calculation energies even if not converged (but still warn!), and use directory name rather than "Label_not_Recognized" if name can't be parsed (more informative). Add test for this
Former-commit-id: ce7c373d652e91ffcb674d5eb7a7e88f8cdb4062
