### 2022-10-18 61ed475
refactor local tests
