+++ tests/test_local.py
@@ -20,10 +20,11 @@ from click.testing import CliRunner
 from doped import vasp_input
 from matplotlib.testing.compare import compare_images
 from monty.serialization import dumpfn
+from pymatgen.analysis.defects.core import StructureMatcher
 from pymatgen.core.structure import Structure
 from pymatgen.io.vasp.inputs import Incar, Kpoints, Poscar
 
-from shakenbreak import input, vasp
+from shakenbreak import input, vasp, cli
 from shakenbreak.cli import snb
 
 _file_path = os.path.dirname(__file__)
@@ -73,8 +74,38 @@ class DistortionLocalTestCase(unittest.TestCase):
             os.path.join(self.VASP_CDTE_DATA_DIR, "CdTe_defects_dict.pickle"), "rb"
         ) as fp:
             self.cdte_defect_dict = pickle.load(fp)
+        # Refactor doped defect dict to dict of Defect() objects
+        self.cdte_defects = {
+            defect_type: [
+                cli.generate_defect_object(defect_dict, self.cdte_defect_dict["bulk"])
+                for defect_dict
+                in self.cdte_defect_dict[defect_type]
+            ] for defect_type in self.cdte_defect_dict.keys() if defect_type != "bulk"
+        }
+        # Use custom names (similar to DOPED/PyCDT)
+        self.cdte_named_defects = {
+            "vacancies": {
+                "vac_1_Cd": self.cdte_defects["vacancies"][0],
+                "vac_2_Te": self.cdte_defects["vacancies"][1],
+            },
+            "substitutions": {
+                "as_1_Cd_on_Te": self.cdte_defects["substitutions"][0],
+                "as_1_Te_on_Cd": self.cdte_defects["substitutions"][1],
+            },
+            "interstitials": {
+                "Int_Cd_1": self.cdte_defects["interstitials"][0],
+                "Int_Cd_2": self.cdte_defects["interstitials"][1],
+                "Int_Cd_3": self.cdte_defects["interstitials"][2],
+                "Int_Te_1": self.cdte_defects["interstitials"][3],
+                "Int_Te_2": self.cdte_defects["interstitials"][4],
+                "Int_Te_3": self.cdte_defects["interstitials"][5],
+            },
+        }
         self.V_Cd_dict = self.cdte_defect_dict["vacancies"][0]
         self.Int_Cd_2_dict = self.cdte_defect_dict["interstitials"][1]
+        # Refactor to Defect() objects
+        self.V_Cd = cli.generate_defect_object(self.V_Cd_dict, self.cdte_defect_dict["bulk"])
+        self.Int_Cd_2 = cli.generate_defect_object(self.Int_Cd_2_dict, self.cdte_defect_dict["bulk"])
 
         self.V_Cd_struc = Structure.from_file(
             os.path.join(self.VASP_CDTE_DATA_DIR, "CdTe_V_Cd_POSCAR")
@@ -233,8 +264,12 @@ class DistortionLocalTestCase(unittest.TestCase):
         for i in os.listdir("."):
             if "distortion_metadata" in i:
                 os.remove(i)
+            if ".png" in i:
+                os.remove(i)
             elif (
                 "Vac_Cd" in i
+                or "v_Cd" in i
+                or "vac_1_Cd" in i
                 or "Int_Cd" in i
                 or "Wally_McDoodle" in i
                 or "pesky_defects" in i
@@ -322,7 +357,7 @@ class DistortionLocalTestCase(unittest.TestCase):
[…]
