+++ tests/test_input.py
@@ -991,10 +991,10 @@ class InputTestCase(unittest.TestCase):
         # and only `Rattled` and `Unperturbed` folders generated for fully-ionised defects
         self.tearDown()
         self.assertFalse(set(self.cdte_defect_folders).issubset(set(os.listdir())))
-        reduced_V_Cd_dict = self.V_Cd_dict.copy()
-        reduced_V_Cd_dict["charges"] = [0, -2]
+        reduced_V_Cd = copy.copy(self.V_Cd)
+        reduced_V_Cd.user_charges = [0, -2]
         dist = input.Distortions(
-            {"vacancies": [reduced_V_Cd_dict]},
+            {"vacancies": {"vac_1_Cd": reduced_V_Cd}},
             local_rattle=False,
         )
         _, distortion_metadata = dist.write_vasp_files(
@@ -1021,11 +1021,11 @@ class InputTestCase(unittest.TestCase):
         self.assertTrue(os.path.exists("vac_1_Cd_-2/Unperturbed"))
 
         # test rattle kwargs:
-        reduced_V_Cd_dict = self.V_Cd_dict.copy()
-        reduced_V_Cd_dict["charges"] = [0]
+        reduced_V_Cd = copy.copy(self.V_Cd)
+        reduced_V_Cd.user_charges = [0]
         rattling_atom_indices = np.arange(0, 31)  # Only rattle Cd
         dist = input.Distortions(
-            {"vacancies": [reduced_V_Cd_dict]},
+            {"vacancies": {"vac_1_Cd": reduced_V_Cd}},
             oxidation_states=oxidation_states,
             bond_distortions=bond_distortions,
             stdev=0.15,
@@ -1055,12 +1055,12 @@ class InputTestCase(unittest.TestCase):
         )
 
         # test other kwargs:
-        reduced_Int_Cd_2_dict = self.Int_Cd_2_dict.copy()
-        reduced_Int_Cd_2_dict["charges"] = [1]
+        reduced_Int_Cd_2 = copy.copy(self.Int_Cd_2)
+        reduced_Int_Cd_2.user_charges = [1,]
 
         with patch("builtins.print") as mock_Int_Cd_2_print:
             dist = input.Distortions(
-                {"interstitials": [reduced_Int_Cd_2_dict]},
+                {"interstitials": {"Int_Cd_2": reduced_Int_Cd_2}},
                 oxidation_states=oxidation_states,
                 distortion_increment=0.25,
                 distorted_elements={"Int_Cd_2": ["Cd"]},
@@ -1080,18 +1080,18 @@ class InputTestCase(unittest.TestCase):
                 },
                 "defects": {
                     "Int_Cd_2": {
-                        "unique_site": reduced_Int_Cd_2_dict[
+                        "unique_site": self.Int_Cd_2_dict[
                             "bulk_supercell_site"
                         ].frac_coords,
                         "charges": {
                             1: {
                                 "num_nearest_neighbours": 4,
                                 "distorted_atoms": [
-                                    (10, "Cd"),
-                                    (22, "Cd"),
-                                    (29, "Cd"),
-                                    (38, "Te"),
-                                ],
+                                    (11, "Cd"),
+                                    (23, "Cd"),
+                                    (30, "Cd"),
+                                    (39, "Te"),
+                                ],  # Defect added at index 0
                                 "distortion_parameters": {
                                     "bond_distortions": [
                                         -0.5,
@@ -1104,8 +1104,8 @@ class InputTestCase(unittest.TestCase):
                                 },
                             },
[…]
