### 2022-10-18 f5d7181
update build_and_test-yml to run test on `pymatgen>2022.8.23_compatibility`
Former-commit-id: 63c9df483ae56d3b2ca7f3932f35aab484731bb9
