+++ .github/workflows/build_and_test.yml
@@ -9,6 +9,7 @@ on:
     branches:
       - main
       - develop
+      - pymatgen>2022.8.23_compatibility
   workflow_dispatch:
 
 jobs:
@@ -46,9 +47,11 @@ jobs:
           done  # Ignore local tests file, which tests INCAR and POTCAR file writing but not possible on GitHub Actions
           pytest --mpl tests/test_shakenbreak.py  # test output plots
           pytest --mpl tests/test_plotting.py  # test output plots
+          # To generate the test plots on GA:
           # pytest --mpl-generate-path=tests/remote_baseline tests/test_plotting.py  # generate output plots
           # pytest --mpl-generate-path=tests/remote_baseline tests/test_shakenbreak.py  # generate output plots
 
+      # Download test plots
       # - name: Archive test plots
       #   uses: actions/upload-artifact@v3
       #   with:
