### 2022-11-16 #4817 @mkhorton
Hi @rkurchin, I've finished my review. This is an excellent, high-quality code, well-motivated and well-documented--I think it's an exemplar for the community, and I would be happy to recommend publication in JOSS.

In terms of the reproducibility checkbox for the data presented in the paper, for full reproducibility, I would have to re-run the DFT calculations myself, but I don't think this is really required: I'm satisfied the code itself does what it claims, and testing the DFT relaxation seems unnecessary.

For the authors @ireaml et al., I would make these some comments:

- Since ShakeNBreak uses the pymatgen `Defect` class, I would make a connection with @jmmshn who is working on [`pymatgen-analysis-defects`](https://github.com/materialsproject/pymatgen-analysis-defects). This seems like a complementary effort and perhaps there'd be some mutual benefit from talking.
- As a user of a code, I find integrated bash scripts (e.g., via `snb.run`) to be cumbersome, since they are doing a similar job to a workflow management system but in a custom way that, in previous experience with other codes that adopt a similar system, usually ends up being quite brittle. I would love to see integration with e.g. atomate2 or AiiDA to make integrating ShakeNBake into existing workflows easier, and make scaling it easier too. I note that atomate2 has a `run_locally` option that could be used too, since I recognise there is benefit to having an option to run locally, or in a debugging context, etc.
- I didn't create issues for these, but only two hiccups, (1) `importlib_metadata` isn't in the `setup.py`, so required an additional install to run the notebook (I'm not sure if this is strictly required either, since this is in the [stdlib](https://docs.python.org/3/library/importlib.metadata.html) as of Python 3.8), and (2) I was not expecting installing this package to also install a font! I think this is harmless, but perhaps worth mentioning in the README, especially since the Montserrat font is distributed under the SIL license, which I believe requires attribution. If we need a longer discussion on either of these points, let's create an issue to discuss.

Overall, really really excellent! The GIF in the docs is especially a nice touch :)

Thanks for your patience while I prepared this review.
