### 2022-11-16 #4817 @jmmshn
Thanks, @mkhorton, we have been talking already and I'm super interested in incorporating this into the defects workflow in atomate2.
