### 2022-11-17 872288f
mention font installation in README and docs
