### 2022-11-25 #4817 @obaica
Hi @rkurchin, I've finished my review. This is an excellent and easy-to-use (e.g., includes a user-friendly command line interface) code that provides new insights for the defect research community. With this tool, I believe it will enable researchers to gain a deeper and more accurate understanding of defect-related properties, and I would highly recommend it for publication in JOSS.

For the authors @ireaml et al., some comments are listed below.
- The authors should provide some specific input files (e.g., `CdTe_Bulk_Supercell_POSCAR`, `CdTe_V_Cd_POSCAR`) that are mentioned in the `[tutorials](https://github.com/SMTG-UCL/ShakeNBreak/tree/main/tutorials)` .
- Interstitial defects are relatively complex compared to vacancy and substitution defects. I suggest that the authors provide a specific example of how to first generate interstitial defects using `pymatgen.analysis.defects` or `doped/PyCDT`, etc., and then further process them with `ShakeNBreak`.
- In the Section `Generate defects with pymatgen.analysis.defects` of the `[Tutorials](https://shakenbreak.readthedocs.io/en/latest/ShakeNBreak_Example_Workflow.html)` , the comment sentence "# Chech defect fractional coordinates" should be "# Check defect fractional coordinates".

Finally, I hope that more researchers in the defect community will use this tool for their research, or even combine  `ShakeNBreak` and `[nonrad](https://github.com/mturiansky/nonrad)` for a more in-depth study of the nature of defects. I will also promote this tool to the researchers around me.

Thank you very much for your patience due to my delay in this review process.
