### 2022-11-25 6bbcebd
fix typo in docs/example notebook
