### 2022-11-25 71f5bc5
update requirements with `importlib_metadata`
