### 2022-11-25 #4817 @ireaml
Hi @mkhorton, thanks for helpful review!

Regarding the comments:

1. We think that refactoring `snb.run` using a workflow manager might not be worthy in our case. Given the initial learning curve/database setup required for `aiida`/`atomate`, this might become a barrier for some users. However, we agree that for high-throughput applications, integration with a workflow manager will be helpful, so we'll add a notebook showing how this can be done.
2. Good points!
    * We have now added `importlib_metadata` as a requirement (commit [71f5bc57](https://github.com/SMTG-UCL/ShakeNBreak/commit/71f5bc5774d19d3c0cf90ca49c5f616ee0fb7ec7)).
    * And a comment after the installation instructions that the Montserrat font is installed and used by default for plotting (commit [e466a83](https://github.com/SMTG-UCL/ShakeNBreak/commit/872288f0b0f45a3904edddd195bdaa003e466a83)). The [FAQ page](https://scripts.sil.org/cms/scripts/page.php?item_id=OFL-FAQ_web#e71fabc0) of the SIL font license states:

      > 1. Can the fonts be included with Free/Libre and Open Source Software collections such as GNU/Linux and BSD distributions and repositories?
      Yes! Fonts licensed under the OFL can be freely included alongside other software under FLOSS (Free/Libre and Open Source Software) licenses. (...)
      > 2. Is any kind of acknowledgement required?
      No. (...)

      So I think we should be fine, but good catch!

Thanks again! :)
