### 2022-11-25 536307a
add notebook showing how to generate interstitials and apply SnB to them
