### 2022-11-25 #4817 @ireaml
Hi @obaica , thanks for the detailed review - good catch on the typo!

1. Input files are provided in the [`tests/data/vasp/CdTe`](https://github.com/SMTG-UCL/ShakeNBreak/tree/main/tests/data/vasp/CdTe) directory
2. We have now added a second example notebook, showing how to generate interstitial defects using `Doped` and apply `ShakeNBreak` to them. (commit [536307a6](https://github.com/SMTG-UCL/ShakeNBreak/commit/536307a6ef02d2d1cdcdcc88c6c2e05e3872c894))
3. Now corrected with commit [6bbcebd0](https://github.com/SMTG-UCL/ShakeNBreak/commit/6bbcebd00217064749860679b5b463d3faba1ba0)

Thanks again! :)
