### 2022-11-28 #4817 @rkurchin
Some minor editorial suggestions:
* line 11: add comma after "site"
* line 12 -> 10: move "however" to the beginning of the sentence – i.e. "However, the standard modeling approach..."
* 16, 20, 27, 48: hyphenate "low-energy"
* 26, 49: hyphenate "energy-lowering"
* 27: add comma after "efficient"
* 29: "automatise" -> "automate"
* 39: unhyphenate "readily applicable"
* 51: add hyphen after "site" i.e. "analyse site- and orbital-decomposed..."

Otherwise, looks good!
