### 2022-11-29 #4817 @ireaml
Hi @rkurchin , thank you for the detailed read and the comments! They have now been implemented.
The version number is [`v22.11.6`](https://github.com/SMTG-UCL/ShakeNBreak/releases/tag/v22.11.6) and the DOI is `10.5281/zenodo.7377173` ([![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.7377173.svg)](https://doi.org/10.5281/zenodo.7377173)).
