### 2022-11-30 #4817 @Kevin-Mattheus-Moerman
@ireaml I am the AEiC for this track, and here to help process this work for acceptance in JOSS. I checked your repository, the paper, and the archive and all seems in order. I have only some minor points on the paper which I hope you can address:

- [ ] Please spell out `UK` as `United Kingdom` for all affiliations
- [ ] You may remove street names and postal codes for all affiliations, simply having `Institute/department, University, City, Country` is sufficient.
