### 2022-12-01 #4817 @ireaml
Hi @rkurchin , thanks for editing!
Just noticed that one of the references was showing the author initial in the text, so fixed this with [3f22b77e](https://github.com/SMTG-UCL/ShakeNBreak/commit/3f22b77ef073ebb74da155fda1827bf6d055e14a). Is it still possible to update the final PDF?
Sorry about this!
