### 2022-12-01 #4817 @xuanxu
@Kevin-Mattheus-Moerman You can regenerate the final pdf reaccepting the paper via `@editorialbot reaccept`
